Refatoração da estrutura do utilities.py transformando em arquivos separados
