--- MODELS/utilities.py
@@ -1,241 +1,0 @@
-from MODELS.excel_file import *
-import chardet, os
-import numpy as np
-import pandas as pd
-
-
-def detectar_encoding(arquivo):
-    with open(arquivo, "rb") as f:
-        resultado = chardet.detect(f.read(1000))
-        return resultado['encoding']
-    
-def detectar_delimitador(arquivo, enconding):
-    with open(arquivo, "r", encoding=enconding) as f:
-        primeira_linha = f.readline()
-        if "\t" in primeira_linha:
-            return "\t"
-        elif ";" in primeira_linha:
-            return ";"
-        elif "," in primeira_linha:
-            return ","
-        
-def convert_csv_to_xlsx(csv_file):
-
-    if csv_file:
-        excel_file = csv_file.replace(".csv",".xlsx")
-
-        # Detectar enconding do arquivo
-        enconding = detectar_encoding(csv_file)
-        print(enconding)
-
-        # Detectar o delimitador do arquivo
-        delimitador = detectar_delimitador(csv_file,enconding)
-        print(delimitador)
-
-        # Ler o CSV com o encodnding e o delimitador
-        try:   
-            df = pd.read_csv(csv_file,delimiter=delimitador,encoding=enconding)
-        except UnicodeDecodeError:
-            print("⚠️ Erro de decodificação com a codificação detectada, tentando com 'latin1'...")
-            df = pd.read_csv(csv_file,delimiter=',',encoding='latin1')
-
-        # Salvar em excel
-        df.to_excel(excel_file,index=False)
-
-    else:
-        pass
-        #print("❌ Nenhum arquivo selecionado")
-
-def remove_duplicates(xlsx_file, type_table):
-
-    if xlsx_file:
-        no_dup_file = os.path.splitext(xlsx_file)[0] + "_no_dup.xlsx"
-        print(f"📂 Arquivo selecionado: {xlsx_file}")
-
-        # Ler a primeira planilha do Excel
-        try:
-            df = pd.read_excel(xlsx_file, sheet_name=0)  # Lê a primeira aba
-        except Exception as e:
-            print(f"⚠️ Erro ao ler o arquivo Excel: {e}")
-            return
-
-        # Remover hyperlinks de todas as colunas de texto
-        for col in df.select_dtypes(include=["object"]).columns:
-            df[col] = df[col].str.replace(r"http\S+", "", regex=True)
-        
-        # Definir as colunas a serem usadas para remoção de duplicatas
-        colunas_para_comparacao_exp = [
-            "university_id",
-            "campus_id", 
-            "name_from_university",
-            "level",
-            "kind",
-            "shift",
-            "enrollment_semester",
-            "max_payments",
-            "full_price"
-            ]
-        colunas_para_comparacao_msp = [
-            "ID do Campus",
-            "Nome do Curso",
-            "Grau",
-            "Modalidade",
-            "Turno",
-            "Duração do Curso",
-            "Mensalidade sem desconto"
-            ]
-
-        # Escolhe as colunas certas conforme o tipo
-        if type_table == "msp":
-            colunas_para_comparacao = colunas_para_comparacao_msp
-        elif type_table == "exp":
-            colunas_para_comparacao = colunas_para_comparacao_exp
-        else:
-            print("❌ Tipo de tabela não reconhecido.")
-            return
-
-        # Cria a coluna de chave única concatenada
-        df['SKU'] = df[colunas_para_comparacao].astype(str).apply(
-            lambda row: '_'.join(row.values), axis=1
-        )
-
-        # Remove duplicatas pela nova coluna
-        df_no_dup = df.drop_duplicates(subset='SKU', keep='first')
-
-        
-        # Remove a coluna auxiliar
-        df_no_dup = df_no_dup.drop(columns=['SKU'])
-
-        # Salvar o DataFrame limpo em um novo arquivo Excel
-        try:
-            df_no_dup.to_excel(no_dup_file, index=False, engine='openpyxl')
-            print(f"✅ Arquivo salvo como: {no_dup_file}")
-            xlsx_file=None
-        except Exception as e:
-            print(f"⚠️ Erro ao salvar o arquivo Excel: {e}")
-
-    else:
-        print("❌ Nenhum arquivo selecionado")
-
-def fix_cities(xlsx_file):
-
-    if xlsx_file:
-        fixed_file = os.path.splitext(xlsx_file)[0] + "_fixed_cities.xlsx"
-        print(f"📂 Arquivo selecionado: {xlsx_file}")
-
-        # Ler a primeira planilha do Excel
-        try:
-            df = pd.read_excel(xlsx_file, sheet_name=0)  # Lê a primeira aba
-            print(df.columns)
-        except Exception as e:
-            print(f"⚠️ Erro ao ler o arquivo Excel: {e}")
-            return
-
-        # Exemplo de lista de cidades
-        lista_cidades={'AC': ['acrelandia', 'assis-brasil', 'brasileia', 'bujari', 'capixaba', 'cruzeiro-do-sul', 'epitaciolandia', 'feijo', 'jordao', 'mancio-lima', 'manoel-urbano', 'marechal-thaumaturgo', 
-        'placido-de-castro', 'porto-walter', 'rio-branco', 'rodrigues-alves', 'santa-rosa-do-purus', 'senador-guiomard', 'sena-madureira', 'tarauaca', 'xapuri', 'porto-acre'], 'AL': ['agua-branca', 'anadia', 'arapiraca', 'atalaia', 'barra-de-santo-antonio', 'barra-de-sao-miguel', 'batalha', 'belem', 'belo-monte', 'boca-da-mata', 'branquinha', 'cacimbinhas', 'cajueiro', 'campestre', 'campo-alegre', 'campo-grande', 'canapi', 'capela', 'carneiros', 'cha-preta', 'coite-do-noia', 'colonia-leopoldina', 'coqueiro-seco', 'coruripe', 'craibas', 'delmiro-gouveia', 'dois-riachos', 'estrela-de-alagoas', 'feira-grande', 'feliz-deserto', 'flexeiras', 'girau-do-ponciano', 'ibateguara', 'igaci', 'igreja-nova', 'inhapi', 'jacare-dos-homens', 'jacuipe', 'japaratinga', 'jaramataia', 'jequia-da-praia', 'joaquim-gomes', 'jundia', 'junqueiro', 'lagoa-da-canoa', 'limoeiro-de-anadia', 'maceio', 'major-isidoro', 'maragogi', 'maravilha', 'marechal-deodoro', 'maribondo', 'mar-vermelho', 'mata-grande', 'matriz-de-camaragibe', 'messias', 'minador-do-negrao', 'monteiropolis', 'murici', 'novo-lino', 'olho-d-agua-das-flores', 'olho-d-agua-do-casado', 'olho-d-agua-grande', 'olivenca', 'ouro-branco', 'palestina', 'palmeira-dos-indios', 'pao-de-acucar', 'pariconha', 'paripueira', 'passo-de-camaragibe', 'paulo-jacinto', 'penedo', 'piacabucu', 'pilar', 'pindoba', 'piranhas', 'poco-das-trincheiras', 'porto-calvo', 'porto-de-pedras', 'porto-real-do-colegio', 'quebrangulo', 'rio-largo', 'roteiro', 'santa-luzia-do-norte', 'santana-do-ipanema', 'santana-do-mundau', 'sao-bras', 'sao-jose-da-laje', 'sao-jose-da-tapera', 'sao-luis-do-quitunde', 'sao-miguel-dos-campos', 'sao-miguel-dos-milagres', 'sao-sebastiao', 'satuba', 'senador-rui-palmeira', 'tanque-d-arca', 'taquarana', 'teotonio-vilela', 'traipu', 'uniao-dos-palmares', 'vicosa'], 'AM': ['alvaraes', 'amatura', 'anama', 'anori', 'apui', 'atalaia-do-norte', 'autazes', 'barcelos', 'barreirinha', 'benjamin-constant', 'beruri', 'boa-vista-do-ramos', 'boca-do-acre', 'borba', 'caapiranga', 'canutama', 'carauari', 'careiro', 'careiro-da-varzea', 'coari', 'codajas', 'eirunepe', 'envira', 'fonte-boa', 'guajara', 'humaita', 'ipixuna', 'iranduba', 'itacoatiara', 'itamarati', 'itapiranga', 'japura', 'jurua', 'jutai', 'labrea', 'manacapuru', 'manaquiri', 'manaus', 'manicore', 'maraa', 'maues', 'nhamunda', 'nova-olinda-do-norte', 'novo-airao', 'novo-aripuana', 'parintins', 'pauini', 'presidente-figueiredo', 'rio-preto-da-eva', 'santa-isabel-do-rio-negro', 'santo-antonio-do-ica', 'sao-gabriel-da-cachoeira', 'sao-paulo-de-olivenca', 'sao-sebastiao-do-uatuma', 'silves', 'tabatinga', 'tapaua', 'tefe', 'tonantins', 'uarini', 'urucara', 'urucurituba'], 'AP': ['serra-do-navio', 'amapa', 'pedra-branca-do-amapari', 'calcoene', 'cutias', 'ferreira-gomes', 'itaubal', 'laranjal-do-jari', 'macapa', 'mazagao', 'oiapoque', 'porto-grande', 'pracuuba', 'santana', 'tartarugalzinho', 'vitoria-do-jari'], 'BA': ['abaira', 'abare', 'acajutiba', 'adustina', 'agua-fria', 'erico-cardoso', 'aiquara', 'alagoinhas', 'alcobaca', 'almadina', 'amargosa', 'amelia-rodrigues', 'america-dourada', 'anage', 'andarai', 'andorinha', 'angical', 'anguera', 'antas', 'antonio-cardoso', 'antonio-goncalves', 'apora', 'apuarema', 'aracatu', 'aracas', 'araci', 'aramari', 'arataca', 'aratuipe', 'aurelino-leal', 'baianopolis', 'baixa-grande', 'banzae', 'barra', 'barra-da-estiva', 'barra-do-choca', 'barra-do-mendes', 'barra-do-rocha', 'barreiras', 'barro-alto', 'barrocas', 'barro-preto', 'belmonte', 'belo-campo', 'biritinga', 'boa-nova', 'boa-vista-do-tupim', 'bom-jesus-da-lapa', 'bom-jesus-da-serra', 'boninal', 'bonito', 'boquira', 'botupora', 'brejoes', 'brejolandia', 'brotas-de-macaubas', 'brumado', 'buerarema', 'buritirama', 'caatiba', 'cabaceiras-do-paraguacu', 'cachoeira', 'cacule', 'caem', 'caetanos', 'caetite', 'cafarnaum', 'cairu', 'caldeirao-grande', 'camacan', 'camacari', 'camamu', 'campo-alegre-de-lourdes', 'campo-formoso', 'canapolis', 'canarana', 'canavieiras', 'candeal', 'candeias', 'candiba', 'candido-sales', 'cansancao', 'canudos', 'capela-do-alto-alegre', 'capim-grosso', 'caraibas', 'caravelas', 'cardeal-da-silva', 'carinhanha', 'casa-nova', 'castro-alves', 'catolandia', 'catu', 'caturama', 'central', 'chorrocho', 'cicero-dantas', 'cipo', 'coaraci', 'cocos', 'conceicao-da-feira', 'conceicao-do-almeida', 'conceicao-do-coite', 'conceicao-do-jacuipe', 'conde', 'condeuba', 'contendas-do-sincora', 'coracao-de-maria', 'cordeiros', 'coribe', 'coronel-joao-sa', 'correntina', 'cotegipe', 'cravolandia', 'crisopolis', 'cristopolis', 'cruz-das-almas', 'curaca', 'dario-meira', 'dias-d-avila', 'dom-basilio', 'dom-macedo-costa', 'elisio-medrado', 'encruzilhada', 'entre-rios', 'esplanada', 'euclides-da-cunha', 'eunapolis', 'fatima', 'feira-da-mata', 'feira-de-santana', 'filadelfia', 'firmino-alves', 'floresta-azul', 'formosa-do-rio-preto', 'gandu', 'gaviao', 'gentio-do-ouro', 'gloria', 'gongogi', 'governador-mangabeira', 'guajeru', 'guanambi', 'guaratinga', 'heliopolis', 'iacu', 'ibiassuce', 'ibicarai', 'ibicoara', 'ibicui', 'ibipeba', 'ibipitanga', 'ibiquera', 'ibirapitanga', 'ibirapua', 'ibirataia', 'ibitiara', 'ibitita', 'ibotirama', 'ichu', 'igapora', 'igrapiuna', 'iguai', 'ilheus', 'inhambupe', 'ipecaeta', 'ipiau', 'ipira', 'ipupiara', 'irajuba', 'iramaia', 'iraquara', 'irara', 'irece', 'itabela', 'itaberaba', 'itabuna', 'itacare', 'itaete', 'itagi', 'itagiba', 'itagimirim', 'itaguacu-da-bahia', 'itaju-do-colonia', 'itajuipe', 'itamaraju', 'itamari', 'itambe', 'itanagra', 'itanhem', 'itaparica', 'itape', 'itapebi', 'itapetinga', 'itapicuru', 'itapitanga', 'itaquara', 'itarantim', 'itatim', 'itirucu', 'itiuba', 'itororo', 'ituacu', 'itubera', 'iuiu', 'jaborandi', 'jacaraci', 'jacobina', 'jaguaquara', 'jaguarari', 'jaguaripe', 'jandaira', 'jequie', 'jeremoabo', 'jiquirica', 'jitauna', 'joao-dourado', 'juazeiro', 'jucurucu', 'jussara', 'jussari', 'jussiape', 'lafaiete-coutinho', 'lagoa-real', 'laje', 'lajedao', 'lajedinho', 'lajedo-do-tabocal', 'lamarao', 'lapao', 'lauro-de-freitas', 'lencois', 'licinio-de-almeida', 'livramento-de-nossa-senhora', 
-        'luis-eduardo-magalhaes', 'macajuba', 'macarani', 'macaubas', 'macurure', 'madre-de-deus', 'maetinga', 'maiquinique', 'mairi', 'malhada', 'malhada-de-pedras', 'manoel-vitorino', 'mansidao', 'maracas', 'maragogipe', 'marau', 'marcionilio-souza', 'mascote', 'mata-de-sao-joao', 'matina', 'medeiros-neto', 'miguel-calmon', 'milagres', 'mirangaba', 'mirante', 'monte-santo', 'morpara', 'morro-do-chapeu', 'mortugaba', 'mucuge', 'mucuri', 'mulungu-do-morro', 'mundo-novo', 'muniz-ferreira', 'muquem-de-sao-francisco', 'muritiba', 'mutuipe', 'nazare', 'nilo-pecanha', 'nordestina', 'nova-canaa', 'nova-fatima', 'nova-ibia', 'nova-itarana', 'nova-redencao', 'nova-soure', 'nova-vicosa', 'novo-horizonte', 'novo-triunfo', 'olindina', 'oliveira-dos-brejinhos', 'ouricangas', 'ourolandia', 'palmas-de-monte-alto', 'palmeiras', 'paramirim', 'paratinga', 'paripiranga', 'pau-brasil', 'paulo-afonso', 'pe-de-serra', 'pedrao', 'pedro-alexandre', 'piata', 'pilao-arcado', 'pindai', 'pindobacu', 'pintadas', 'pirai-do-norte', 'piripa', 'piritiba', 'planaltino', 'planalto', 'pocoes', 'pojuca', 'ponto-novo', 'porto-seguro', 'potiragua', 'prado', 'presidente-dutra', 'presidente-janio-quadros', 'presidente-tancredo-neves', 'queimadas', 'quijingue', 'quixabeira', 'rafael-jambeiro', 'remanso', 'retirolandia', 'riachao-das-neves', 'riachao-do-jacuipe', 'riacho-de-santana', 'ribeira-do-amparo', 'ribeira-do-pombal', 'ribeirao-do-largo', 'rio-de-contas', 'rio-do-antonio', 'rio-do-pires', 'rio-real', 'rodelas', 'ruy-barbosa', 'salinas-da-margarida', 'salvador', 'santa-barbara', 'santa-brigida', 'santa-cruz-cabralia', 'santa-cruz-da-vitoria', 'santa-ines', 'santaluz', 'santa-luzia', 'santa-maria-da-vitoria', 'santana', 'santanopolis', 'santa-rita-de-cassia', 'santa-teresinha', 'santo-amaro', 'santo-antonio-de-jesus', 'santo-estevao', 'sao-desiderio', 'sao-domingos', 'sao-felix', 'sao-felix-do-coribe', 'sao-felipe', 'sao-francisco-do-conde', 'sao-gabriel', 'sao-goncalo-dos-campos', 'sao-jose-da-vitoria', 'sao-jose-do-jacuipe', 'sao-miguel-das-matas', 'sao-sebastiao-do-passe', 'sapeacu', 'satiro-dias', 'saubara', 'saude', 'seabra', 'sebastiao-laranjeiras', 'senhor-do-bonfim', 'serra-do-ramalho', 'sento-se', 'serra-dourada', 'serra-preta', 'serrinha', 'serrolandia', 'simoes-filho', 'sitio-do-mato', 'sitio-do-quinto', 'sobradinho', 'souto-soares', 'tabocas-do-brejo-velho', 'tanhacu', 'tanque-novo', 'tanquinho', 'taperoa', 'tapiramuta', 'teixeira-de-freitas', 'teodoro-sampaio', 'teofilandia', 'teolandia', 'terra-nova', 'tremedal', 'tucano', 'uaua', 'ubaira', 'ubaitaba', 'ubata', 'uibai', 'umburanas', 'una', 'urandi', 'urucuca', 'utinga', 'valenca', 'valente', 'varzea-da-roca', 'varzea-do-poco', 'varzea-nova', 'varzedo', 'vera-cruz', 'vereda', 'vitoria-da-conquista', 'wagner', 'wanderley', 'wenceslau-guimaraes', 'xique-xique'], 'CE': ['abaiara', 'acarape', 'acarau', 'acopiara', 'aiuaba', 'alcantaras', 'altaneira', 'alto-santo', 'amontada', 'antonina-do-norte', 'apuiares', 'aquiraz', 'aracati', 'aracoiaba', 'ararenda', 'araripe', 'aratuba', 'arneiroz', 'assare', 'aurora', 'baixio', 'banabuiu', 'barbalha', 'barreira', 'barro', 'barroquinha', 'baturite', 'beberibe', 'bela-cruz', 'boa-viagem', 'brejo-santo', 'camocim', 'campos-sales', 'caninde', 'capistrano', 'caridade', 'carire', 'caririacu', 'carius', 'carnaubal', 'cascavel', 'catarina', 'catunda', 'caucaia', 'cedro', 'chaval', 'choro', 'chorozinho', 'coreau', 'crateus', 'crato', 'croata', 'cruz', 'deputado-irapuan-pinheiro', 'erere', 'eusebio', 'farias-brito', 'forquilha', 'fortaleza', 'fortim', 'frecheirinha', 'general-sampaio', 'graca', 'granja', 'granjeiro', 'groairas', 'guaiuba', 'guaraciaba-do-norte', 'guaramiranga', 'hidrolandia', 'horizonte', 'ibaretama', 'ibiapina', 'ibicuitinga', 'icapui', 'ico', 'iguatu', 'independencia', 'ipaporanga', 'ipaumirim', 'ipu', 'ipueiras', 'iracema', 'iraucuba', 'itaicaba', 'itaitinga', 'itapage', 'itapipoca', 'itapiuna', 'itarema', 'itatira', 'jaguaretama', 'jaguaribara', 'jaguaribe', 'jaguaruana', 'jardim', 'jati', 'jijoca-de-jericoacoara', 'juazeiro-do-norte', 'jucas', 'lavras-da-mangabeira', 'limoeiro-do-norte', 'madalena', 'maracanau', 'maranguape', 'marco', 'martinopole', 'massape', 'mauriti', 'meruoca', 'milagres', 'milha', 'miraima', 'missao-velha', 'mombaca', 'monsenhor-tabosa', 'morada-nova', 'moraujo', 'morrinhos', 'mucambo', 'mulungu', 'nova-olinda', 'nova-russas', 'novo-oriente', 'ocara', 'oros', 'pacajus', 
-        'pacatuba', 'pacoti', 'pacuja', 'palhano', 'palmacia', 'paracuru', 'paraipaba', 'parambu', 'paramoti', 'pedra-branca', 'penaforte', 'pentecoste', 'pereiro', 'pindoretama', 'piquet-carneiro', 'pires-ferreira', 'poranga', 'porteiras', 'potengi', 'potiretama', 'quiterianopolis', 'quixada', 'quixelo', 'quixeramobim', 'quixere', 'redencao', 'reriutaba', 'russas', 'saboeiro', 'salitre', 'santana-do-acarau', 'santana-do-cariri', 'santa-quiteria', 'sao-benedito', 'sao-goncalo-do-amarante', 'sao-joao-do-jaguaribe', 'sao-luis-do-curu', 'senador-pompeu', 'senador-sa', 'sobral', 'solonopole', 'tabuleiro-do-norte', 'tamboril', 'tarrafas', 'taua', 'tejucuoca', 'tiangua', 'trairi', 'tururu', 'ubajara', 'umari', 'umirim', 'uruburetama', 'uruoca', 'varjota', 'varzea-alegre', 'vicosa-do-ceara'], 'DF': ['santa-maria', 'brazlandia', 'ceilandia', 'samambaia', 'planaltina', 'taguatinga', 'brasilia', 'sobradinho', 'sao-sebastiao', 'aguas-claras', 'gama'], 'ES': ['afonso-claudio', 'aguia-branca', 'agua-doce-do-norte', 'alegre', 'alfredo-chaves', 'alto-rio-novo', 'anchieta', 'apiaca', 'aracruz', 'atilio-vivacqua', 'baixo-guandu', 'barra-de-sao-francisco', 'boa-esperanca', 'bom-jesus-do-norte', 'brejetuba', 'cachoeiro-de-itapemirim', 'cariacica', 'castelo', 'colatina', 'conceicao-da-barra', 'conceicao-do-castelo', 'divino-de-sao-lourenco', 'domingos-martins', 'dores-do-rio-preto', 'ecoporanga', 'fundao', 'governador-lindenberg', 'guacui', 'guarapari', 'ibatiba', 'ibiracu', 'ibitirama', 'iconha', 'irupi', 'itaguacu', 'itapemirim', 'itarana', 'iuna', 'jaguare', 'jeronimo-monteiro', 'joao-neiva', 'laranja-da-terra', 'linhares', 'mantenopolis', 'marataizes', 'marechal-floriano', 'marilandia', 'mimoso-do-sul', 'montanha', 'mucurici', 'muniz-freire', 'muqui', 'nova-venecia', 'pancas', 'pedro-canario', 'pinheiros', 'piuma', 'ponto-belo', 'presidente-kennedy', 'rio-bananal', 'rio-novo-do-sul', 'santa-leopoldina', 'santa-maria-de-jetiba', 'santa-teresa', 'sao-domingos-do-norte', 'sao-gabriel-da-palha', 'sao-jose-do-calcado', 'sao-mateus', 'sao-roque-do-canaa', 'serra', 'sooretama', 'vargem-alta', 'venda-nova-do-imigrante', 'viana', 'vila-pavao', 'vila-valerio', 'vila-velha', 'vitoria'], 'GO': ['abadia-de-goias', 'abadiania', 'acreuna', 'adelandia', 'agua-fria-de-goias', 'agua-limpa', 'aguas-lindas-de-goias', 'alexania', 'aloandia', 'alto-horizonte', 'alto-paraiso-de-goias', 'alvorada-do-norte', 'amaralina', 'americano-do-brasil', 'amorinopolis', 'anapolis', 'anhanguera', 'anicuns', 'aparecida-de-goiania', 'aparecida-do-rio-doce', 'apore', 'aracu', 'aragarcas', 'aragoiania', 'araguapaz', 'arenopolis', 'aruana', 'aurilandia', 'avelinopolis', 'baliza', 'barro-alto', 'bela-vista-de-goias', 'bom-jardim-de-goias', 'bom-jesus-de-goias', 'bonfinopolis', 'bonopolis', 'brazabrantes', 'britania', 'buriti-alegre', 'buriti-de-goias', 'buritinopolis', 'cabeceiras', 'cachoeira-alta', 'cachoeira-de-goias', 'cachoeira-dourada', 'cacu', 'caiaponia', 'caldas-novas', 'caldazinha', 'campestre-de-goias', 'campinacu', 'campinorte', 'campo-alegre-de-goias', 'campo-limpo-de-goias', 'campos-belos', 'campos-verdes', 'carmo-do-rio-verde', 'castelandia', 'catalao', 'caturai', 'cavalcante', 'ceres', 'cezarina', 'chapadao-do-ceu', 'cidade-ocidental', 'cocalzinho-de-goias', 'colinas-do-sul', 'corrego-do-ouro', 'corumba-de-goias', 'corumbaiba', 'cristalina', 'cristianopolis', 'crixas', 'crominia', 'cumari', 'damianopolis', 'damolandia', 'davinopolis', 'diorama', 'doverlandia', 'edealina', 'edeia', 'estrela-do-norte', 
-        'faina', 'fazenda-nova', 'firminopolis', 'flores-de-goias', 'formosa', 'formoso', 'gameleira-de-goias', 'divinopolis-de-goias', 'goianapolis', 'goiandira', 'goianesia', 'goiania', 'goianira', 'goias', 'goiatuba', 'gouvelandia', 'guapo', 'guaraita', 'guarani-de-goias', 'guarinos', 'heitorai', 'hidrolandia', 'hidrolina', 'iaciara', 'inaciolandia', 'indiara', 'inhumas', 
-        'ipameri', 'ipiranga-de-goias', 'ipora', 'israelandia', 'itaberai', 'itaguari', 'itaguaru', 'itaja', 'itapaci', 'itapirapua', 'itapuranga', 'itaruma', 'itaucu', 'itumbiara', 'ivolandia', 'jandaia', 'jaragua', 'jatai', 'jaupaci', 'jesupolis', 'joviania', 'jussara', 'lagoa-santa', 'leopoldo-de-bulhoes', 'luziania', 'mairipotaba', 'mambai', 'mara-rosa', 'marzagao', 'matrincha', 'maurilandia', 'mimoso-de-goias', 'minacu', 'mineiros', 'moipora', 'monte-alegre-de-goias', 'montes-claros-de-goias', 'montividiu', 'montividiu-do-norte', 'morrinhos', 'morro-agudo-de-goias', 'mossamedes', 'mozarlandia', 'mundo-novo', 'mutunopolis', 'nazario', 'neropolis', 'niquelandia', 'nova-america', 'nova-aurora', 'nova-crixas', 'nova-gloria', 'nova-iguacu-de-goias', 'nova-roma', 'nova-veneza', 'novo-brasil', 'novo-gama', 'novo-planalto', 'orizona', 'ouro-verde-de-goias', 'ouvidor', 'padre-bernardo', 'palestina-de-goias', 'palmeiras-de-goias', 'palmelo', 'palminopolis', 'panama', 'paranaiguara', 'parauna', 'perolandia', 'petrolina-de-goias', 'pilar-de-goias', 'piracanjuba', 'piranhas', 'pirenopolis', 'pires-do-rio', 'planaltina', 'pontalina', 'porangatu', 'porteirao', 'portelandia', 'posse', 'professor-jamil', 'quirinopolis', 'rialma', 'rianapolis', 'rio-quente', 'rio-verde', 'rubiataba', 'sanclerlandia', 'santa-barbara-de-goias', 'santa-cruz-de-goias', 'santa-fe-de-goias', 'santa-helena-de-goias', 'santa-isabel', 'santa-rita-do-araguaia', 'santa-rita-do-novo-destino', 'santa-rosa-de-goias', 'santa-tereza-de-goias', 'santa-terezinha-de-goias', 'santo-antonio-da-barra', 'santo-antonio-de-goias', 'santo-antonio-do-descoberto', 'sao-domingos', 'sao-francisco-de-goias', 'sao-joao-d-alianca', 'sao-joao-da-parauna', 'sao-luis-de-montes-belos', 'sao-luiz-do-norte', 'sao-miguel-do-araguaia', 'sao-miguel-do-passa-quatro', 'sao-patricio', 'sao-simao', 'senador-canedo', 'serranopolis', 'silvania', 'simolandia', 'sitio-d-abadia', 'taquaral-de-goias', 'teresina-de-goias', 'terezopolis-de-goias', 'tres-ranchos', 'trindade', 'trombas', 'turvania', 'turvelandia', 'uirapuru', 'uruacu', 'uruana', 'urutai', 'valparaiso-de-goias', 'varjao', 'vianopolis', 'vicentinopolis', 'vila-boa', 'vila-propicio'], 'MA': ['arame', 'acailandia', 'afonso-cunha', 'agua-doce-do-maranhao', 'alcantara', 'aldeias-altas', 'altamira-do-maranhao', 'alto-alegre-do-maranhao', 'alto-alegre-do-pindare', 'alto-parnaiba', 'amapa-do-maranhao', 'amarante-do-maranhao', 'anajatuba', 'anapurus', 'apicum-acu', 'araguana', 'araioses', 'arari', 'axixa', 'bacabal', 'bacabeira', 'bacuri', 'bacurituba', 'balsas', 'barao-de-grajau', 'barra-do-corda', 'barreirinhas', 'belagua', 'bela-vista-do-maranhao', 'benedito-leite', 'bequimao', 'bernardo-do-mearim', 'boa-vista-do-gurupi', 'bom-jardim', 'bom-jesus-das-selvas', 'bom-lugar', 'brejo', 'brejo-de-areia', 'buriti', 'buriti-bravo', 'buriticupu', 'buritirana', 'cachoeira-grande', 'cajapio', 'cajari', 'campestre-do-maranhao', 'candido-mendes', 'cantanhede', 'capinzal-do-norte', 'carolina', 'carutapera', 'caxias', 'cedral', 'central-do-maranhao', 'centro-do-guilherme', 'centro-novo-do-maranhao', 'chapadinha', 'cidelandia', 'codo', 
-        'coelho-neto', 'colinas', 'conceicao-do-lago-acu', 'coroata', 'cururupu', 'davinopolis', 'dom-pedro', 'duque-bacelar', 'esperantinopolis', 'estreito', 'feira-nova-do-maranhao', 'fernando-falcao', 'formosa-da-serra-negra', 'fortaleza-dos-nogueiras', 'fortuna', 'godofredo-viana', 'goncalves-dias', 'governador-archer', 'governador-edison-lobao', 'governador-eugenio-barros', 'governador-luiz-rocha', 'governador-newton-bello', 'governador-nunes-freire', 'graca-aranha', 'grajau', 'guimaraes', 'humberto-de-campos', 'icatu', 'igarape-do-meio', 'igarape-grande', 'imperatriz', 'itaipava-do-grajau', 'itapecuru-mirim', 'itinga-do-maranhao', 'jatoba', 'jenipapo-dos-vieiras', 'joao-lisboa', 'joselandia', 'junco-do-maranhao', 'lago-da-pedra', 'lago-do-junco', 'lago-verde', 'lagoa-do-mato', 'lago-dos-rodrigues', 'lagoa-grande-do-maranhao', 'lajeado-novo', 'lima-campos', 'loreto', 'luis-domingues', 'magalhaes-de-almeida', 'maracacume', 'maraja-do-sena', 'maranhaozinho', 'mata-roma', 'matinha', 'matoes', 'matoes-do-norte', 'milagres-do-maranhao', 'mirador', 'miranda-do-norte', 'mirinzal', 'moncao', 'montes-altos', 'morros', 'nina-rodrigues', 'nova-colinas', 'nova-iorque', 'nova-olinda-do-maranhao', 'olho-d-agua-das-cunhas', 'olinda-nova-do-maranhao', 'paco-do-lumiar', 'palmeirandia', 'paraibano', 'parnarama', 'passagem-franca', 'pastos-bons', 'paulino-neves', 'paulo-ramos', 'pedreiras', 'pedro-do-rosario', 'penalva', 'peri-mirim', 'peritoro', 'pindare-mirim', 'pinheiro', 'pio-xii', 'pirapemas', 'pocao-de-pedras', 'porto-franco', 'porto-rico-do-maranhao', 'presidente-dutra', 'presidente-juscelino', 'presidente-medici', 'presidente-sarney', 'presidente-vargas', 'primeira-cruz', 'raposa', 'riachao', 'ribamar-fiquene', 'rosario', 'sambaiba', 'santa-filomena-do-maranhao', 'santa-helena', 'santa-ines', 'santa-luzia', 'santa-luzia-do-parua', 'santa-quiteria-do-maranhao', 'santa-rita', 'santana-do-maranhao', 'santo-amaro-do-maranhao', 'santo-antonio-dos-lopes', 'sao-benedito-do-rio-preto', 'sao-bento', 'sao-bernardo', 'sao-domingos-do-azeitao', 'sao-domingos-do-maranhao', 'sao-felix-de-balsas', 'sao-francisco-do-brejao', 'sao-francisco-do-maranhao', 'sao-joao-batista', 'sao-joao-do-caru', 'sao-joao-do-paraiso', 'sao-joao-do-soter', 'sao-joao-dos-patos', 'sao-jose-de-ribamar', 'sao-jose-dos-basilios', 'sao-luis', 'sao-luis-gonzaga-do-maranhao', 'sao-mateus-do-maranhao', 'sao-pedro-da-agua-branca', 'sao-pedro-dos-crentes', 'sao-raimundo-das-mangabeiras', 'sao-raimundo-do-doca-bezerra', 'sao-roberto', 'sao-vicente-ferrer', 'satubinha', 'senador-alexandre-costa', 'senador-la-rocque', 'serrano-do-maranhao', 'sitio-novo', 'sucupira-do-norte', 'sucupira-do-riachao', 'tasso-fragoso', 'timbiras', 'timon', 'trizidela-do-vale', 'tufilandia', 'tuntum', 'turiacu', 'turilandia', 'tutoia', 'urbano-santos', 'vargem-grande', 'viana', 'vila-nova-dos-martirios', 'vitoria-do-mearim', 'vitorino-freire', 'ze-doca'], 'MG': ['serro', 'abadia-dos-dourados', 'abaete', 'abre-campo', 'acaiaca', 'acucena', 'agua-boa', 'agua-comprida', 'aguanil', 'aguas-formosas', 'aguas-vermelhas', 'aimores', 'aiuruoca', 'alagoa', 'albertina', 'alem-paraiba', 'alfenas', 'bugre', 'alfredo-vasconcelos', 'almenara', 'alpercata', 'alpinopolis', 'alterosa', 'alto-caparao', 'alto-rio-doce', 'alvarenga', 'alvinopolis', 'alvorada-de-minas', 'amparo-do-serra', 'andradas', 'cachoeira-de-pajeu', 'andrelandia', 'angelandia', 'antonio-carlos', 'antonio-dias', 'antonio-prado-de-minas', 'aracai', 'aracitaba', 'aracuai', 'araguari', 'arantina', 'araponga', 'arapora', 'arapua', 'araujos', 'araxa', 'arceburgo', 'arcos', 'areado', 'argirita', 'aricanduva', 'arinos', 'astolfo-dutra', 'ataleia', 'augusto-de-lima', 'baependi', 'baldim', 'bambui', 'bandeira', 'bandeira-do-sul', 'barao-de-cocais', 'barao-de-monte-alto', 'barbacena', 'barra-longa', 'barroso', 'bela-vista-de-minas', 'belmiro-braga', 'belo-horizonte', 'belo-oriente', 'belo-vale', 'berilo', 'bertopolis', 'berizal', 'betim', 'bias-fortes', 'bicas', 'biquinhas', 'boa-esperanca', 'bocaina-de-minas', 'bocaiuva', 'bom-despacho', 'bom-jardim-de-minas', 'bom-jesus-da-penha', 'bom-jesus-do-amparo', 'bom-jesus-do-galho', 'bom-repouso', 'bom-sucesso', 'bonfim', 'bonfinopolis-de-minas', 'bonito-de-minas', 'borda-da-mata', 'botelhos', 'botumirim', 'brasilandia-de-minas', 'brasilia-de-minas', 'bras-pires', 'braunas', 'brasopolis', 'brumadinho', 'bueno-brandao', 'buenopolis', 'buritis', 'buritizeiro', 'cabeceira-grande', 'cabo-verde', 'cachoeira-da-prata', 'cachoeira-de-minas', 'cachoeira-dourada', 'caetanopolis', 'caete', 'caiana', 'cajuri', 'caldas', 'camacho', 'camanducaia', 'cambui', 'cambuquira', 'campanario', 'campanha', 'campestre', 'campina-verde', 'campo-azul', 'campo-belo', 'campo-do-meio', 'campo-florido', 'campos-altos', 'campos-gerais', 'canaa', 'canapolis', 'cana-verde', 'candeias', 'cantagalo', 'caparao', 'capela-nova', 'capelinha', 'capetinga', 'capim-branco', 'capinopolis', 'capitao-andrade', 'capitao-eneas', 'capitolio', 'caputira', 'carai', 'caranaiba', 'carandai', 'carangola', 'caratinga', 'carbonita', 'careacu', 'carlos-chagas', 'carmesia', 'carmo-da-cachoeira', 'carmo-da-mata', 'carmo-de-minas', 'carmo-do-cajuru', 'carmo-do-paranaiba', 'carmo-do-rio-claro', 'carmopolis-de-minas', 'carneirinho', 'carrancas', 'carvalhopolis', 'carvalhos', 'casa-grande', 'cascalho-rico', 'cassia', 'conceicao-da-barra-de-minas', 'cataguases', 'catas-altas', 'catas-altas-da-noruega', 'catuji', 'catuti', 'caxambu', 'cedro-do-abaete', 'central-de-minas', 'centralina', 'chacara', 'chale', 'chapada-do-norte', 'chapada-gaucha', 'chiador', 'cipotanea', 'claraval', 'claro-dos-pocoes', 'claudio', 'coimbra', 'coluna', 'comendador-gomes', 'comercinho', 'conceicao-da-aparecida', 'conceicao-das-pedras', 'conceicao-das-alagoas', 'conceicao-de-ipanema', 'conceicao-do-mato-dentro', 'conceicao-do-para', 'conceicao-do-rio-verde', 'conceicao-dos-ouros', 'conego-marinho', 
-        'confins', 'congonhal', 'congonhas', 'congonhas-do-norte', 'conquista', 'conselheiro-lafaiete', 'conselheiro-pena', 'consolacao', 'contagem', 'coqueiral', 'coracao-de-jesus', 'cordisburgo', 'cordislandia', 'corinto', 'coroaci', 'coromandel', 'coronel-fabriciano', 'coronel-murta', 'coronel-pacheco', 'coronel-xavier-chaves', 'corrego-danta', 'corrego-do-bom-jesus', 'corrego-fundo', 'corrego-novo', 'couto-de-magalhaes-de-minas', 'crisolita', 'cristais', 'cristalia', 'cristiano-otoni', 'cristina', 'crucilandia', 'cruzeiro-da-fortaleza', 'cruzilia', 'cuparaque', 'curral-de-dentro', 'curvelo', 'datas', 'delfim-moreira', 'delfinopolis', 'delta', 'descoberto', 'desterro-de-entre-rios', 'desterro-do-melo', 'diamantina', 'diogo-de-vasconcelos', 'dionisio', 'divinesia', 'divino', 'divino-das-laranjeiras', 'divinolandia-de-minas', 'divinopolis', 'divisa-alegre', 'divisa-nova', 'divisopolis', 'dom-bosco', 'dom-cavati', 'dom-joaquim', 'dom-silverio', 'dom-vicoso', 'dona-eusebia', 'dores-de-campos', 'dores-de-guanhaes', 'dores-do-indaia', 'dores-do-turvo', 'doresopolis', 'douradoquara', 'durande', 'eloi-mendes', 'engenheiro-caldas', 'engenheiro-navarro', 'entre-folhas', 'entre-rios-de-minas', 'ervalia', 'esmeraldas', 'espera-feliz', 'espinosa', 'espirito-santo-do-dourado', 'estiva', 'estrela-dalva', 'estrela-do-indaia', 'estrela-do-sul', 'eugenopolis', 'ewbank-da-camara', 'extrema', 'fama', 'faria-lemos', 'felicio-dos-santos', 'sao-goncalo-do-rio-preto', 'felisburgo', 'felixlandia', 'fernandes-tourinho', 'ferros', 'fervedouro', 'florestal', 'formiga', 'formoso', 'fortaleza-de-minas', 'fortuna-de-minas', 'francisco-badaro', 'francisco-dumont', 'francisco-sa', 'franciscopolis', 'frei-gaspar', 'frei-inocencio', 'frei-lagonegro', 'fronteira', 'fronteira-dos-vales', 'fruta-de-leite', 'frutal', 'funilandia', 'galileia', 'gameleiras', 'glaucilandia', 'goiabeira', 'goiana', 'goncalves', 'gonzaga', 'gouveia', 'governador-valadares', 'grao-mogol', 'grupiara', 'guanhaes', 'guape', 'guaraciaba', 'guaraciama', 'guaranesia', 'guarani', 'guarara', 'guarda-mor', 'guaxupe', 'guidoval', 'guimarania', 'guiricema', 'gurinhata', 'heliodora', 'iapu', 'ibertioga', 'ibia', 'ibiai', 'ibiracatu', 'ibiraci', 'ibirite', 'ibitiura-de-minas', 'ibituruna', 'icarai-de-minas', 'igarape', 'igaratinga', 'iguatama', 'ijaci', 'ilicinea', 'imbe-de-minas', 'inconfidentes', 'indaiabira', 'indianopolis', 'ingai', 'inhapim', 'inhauma', 'inimutaba', 'ipaba', 'ipanema', 'ipatinga', 'ipiacu', 'ipuiuna', 'irai-de-minas', 'itabira', 'itabirinha', 'itabirito', 'itacambira', 'itacarambi', 'itaguara', 'itaipe', 'itajuba', 'itamarandiba', 'itamarati-de-minas', 'itambacuri', 'itambe-do-mato-dentro', 'itamogi', 'itamonte', 'itanhandu', 'itanhomi', 'itaobim', 'itapagipe', 'itapecerica', 'itapeva', 'itatiaiucu', 'itau-de-minas', 'itauna', 'itaverava', 'itinga', 'itueta', 'ituiutaba', 'itumirim', 'iturama', 'itutinga', 'jaboticatubas', 'jacinto', 'jacui', 'jacutinga', 'jaguaracu', 'jaiba', 'jampruca', 'janauba', 'januaria', 'japaraiba', 'japonvar', 'jeceaba', 'jenipapo-de-minas', 'jequeri', 'jequitai', 'jequitiba', 'jequitinhonha', 'jesuania', 'joaima', 'joanesia', 'joao-monlevade', 'joao-pinheiro', 'joaquim-felicio', 'jordania', 'jose-goncalves-de-minas', 'jose-raydan', 'josenopolis', 'nova-uniao', 'juatuba', 'juiz-de-fora', 'juramento', 
-        'juruaia', 'juvenilia', 'ladainha', 'lagamar', 'lagoa-da-prata', 'lagoa-dos-patos', 'lagoa-dourada', 'lagoa-formosa', 'lagoa-grande', 'lagoa-santa', 'lajinha', 'lambari', 'lamim', 'laranjal', 'lassance', 'lavras', 'leandro-ferreira', 'leme-do-prado', 'leopoldina', 'liberdade', 'lima-duarte', 'limeira-do-oeste', 'lontra', 'luisburgo', 'luislandia', 'luminarias', 'luz', 'machacalis', 'machado', 'madre-de-deus-de-minas', 'malacacheta', 'mamonas', 'manga', 'manhuacu', 'manhumirim', 'mantena', 'maravilhas', 'mar-de-espanha', 'maria-da-fe', 'mariana', 'marilac', 'mario-campos', 'maripa-de-minas', 'marlieria', 'marmelopolis', 'martinho-campos', 'martins-soares', 'mata-verde', 'materlandia', 'mateus-leme', 'matias-barbosa', 'matias-cardoso', 'matipo', 'mato-verde', 'matozinhos', 'matutina', 'medeiros', 'medina', 'mendes-pimentel', 'merces', 'mesquita', 'minas-novas', 'minduri', 'mirabela', 'miradouro', 'mirai', 'miravania', 'moeda', 'moema', 'monjolos', 'monsenhor-paulo', 'montalvania', 'monte-alegre-de-minas', 'monte-azul', 'monte-belo', 'monte-carmelo', 'monte-formoso', 'monte-santo-de-minas', 'montes-claros', 'monte-siao', 'montezuma', 'morada-nova-de-minas', 'morro-da-garca', 'morro-do-pilar', 'munhoz', 'muriae', 'mutum', 'muzambinho', 'nacip-raydan', 'nanuque', 'naque', 'natalandia', 'natercia', 'nazareno', 'nepomuceno', 'ninheira', 'nova-belem', 'nova-era', 'nova-lima', 'nova-modica', 'nova-ponte', 'nova-porteirinha', 'nova-resende', 'nova-serrana', 'novo-cruzeiro', 'novo-oriente-de-minas', 'novorizonte', 'olaria', 'olhos-d-agua', 'olimpio-noronha', 'oliveira', 'oliveira-fortes', 'onca-de-pitangui', 'oratorios', 'orizania', 'ouro-branco', 
-        'ouro-fino', 'ouro-preto', 'ouro-verde-de-minas', 'padre-carvalho', 'padre-paraiso', 'paineiras', 'pains', 'pai-pedro', 'paiva', 'palma', 'palmopolis', 'papagaios', 'paracatu', 'para-de-minas', 'paraguacu', 'paraisopolis', 'paraopeba', 'passabem', 'passa-quatro', 'passa-tempo', 'passa-vinte', 'passos', 'patis', 'patos-de-minas', 'patrocinio', 'patrocinio-do-muriae', 'paula-candido', 'paulistas', 'pavao', 'pecanha', 'pedra-azul', 'pedra-bonita', 'pedra-do-anta', 'pedra-do-indaia', 'pedra-dourada', 'pedralva', 'pedras-de-maria-da-cruz', 'pedrinopolis', 'pedro-leopoldo', 'pedro-teixeira', 'pequeri', 'pequi', 'perdigao', 'perdizes', 'perdoes', 'periquito', 'pescador', 'piau', 'piedade-de-caratinga', 'piedade-de-ponte-nova', 'piedade-do-rio-grande', 'piedade-dos-gerais', 'pimenta', 'pingo-d-agua', 'pintopolis', 'piracema', 'pirajuba', 'piranga', 'pirangucu', 'piranguinho', 'pirapetinga', 'pirapora', 'pirauba', 'pitangui', 'piumhi', 'planura', 'poco-fundo', 'pocos-de-caldas', 'pocrane', 'pompeu', 'ponte-nova', 'ponto-chique', 'ponto-dos-volantes', 'porteirinha', 'porto-firme', 'pote', 'pouso-alegre', 'pouso-alto', 'prados', 'prata', 'pratapolis', 'pratinha', 'presidente-bernardes', 'presidente-juscelino', 'presidente-kubitschek', 'presidente-olegario', 'alto-jequitiba', 'prudente-de-morais', 'quartel-geral', 'queluzito', 'raposos', 'raul-soares', 'recreio', 'reduto', 'resende-costa', 'resplendor', 'ressaquinha', 'riachinho', 'riacho-dos-machados', 'ribeirao-das-neves', 'ribeirao-vermelho', 'rio-acima', 'rio-casca', 'rio-doce', 'rio-do-prado', 'rio-espera', 'rio-manso', 'rio-novo', 'rio-paranaiba', 'rio-pardo-de-minas', 'rio-piracicaba', 'rio-pomba', 'rio-preto', 'rio-vermelho', 'ritapolis', 'rochedo-de-minas', 'rodeiro', 'romaria', 'rosario-da-limeira', 'rubelita', 'rubim', 'sabara', 'sabinopolis', 'sacramento', 'salinas', 'salto-da-divisa', 'santa-barbara', 'santa-barbara-do-leste', 'santa-barbara-do-monte-verde', 'santa-barbara-do-tugurio', 'santa-cruz-de-minas', 'santa-cruz-de-salinas', 'santa-cruz-do-escalvado', 'santa-efigenia-de-minas', 'santa-fe-de-minas', 'santa-helena-de-minas', 'santa-juliana', 'santa-luzia', 'santa-margarida', 'santa-maria-de-itabira', 'santa-maria-do-salto', 
-        'santa-maria-do-suacui', 'santana-da-vargem', 'santana-de-cataguases', 'santana-de-pirapama', 'santana-do-deserto', 'santana-do-garambeu', 'santana-do-jacare', 'santana-do-manhuacu', 'santana-do-paraiso', 'santana-do-riacho', 'santana-dos-montes', 'santa-rita-de-caldas', 'santa-rita-de-jacutinga', 'santa-rita-de-minas', 'santa-rita-de-ibitipoca', 'santa-rita-do-itueto', 'santa-rita-do-sapucai', 'santa-rosa-da-serra', 'santa-vitoria', 'santo-antonio-do-amparo', 'santo-antonio-do-aventureiro', 'santo-antonio-do-grama', 'santo-antonio-do-itambe', 'santo-antonio-do-jacinto', 'santo-antonio-do-monte', 'santo-antonio-do-retiro', 'santo-antonio-do-rio-abaixo', 'santo-hipolito', 'santos-dumont', 'sao-bento-abade', 'sao-bras-do-suacui', 'sao-domingos-das-dores', 'sao-domingos-do-prata', 'sao-felix-de-minas', 'sao-francisco', 'sao-francisco-de-paula', 'sao-francisco-de-sales', 'sao-francisco-do-gloria', 'sao-geraldo', 'sao-geraldo-da-piedade', 'sao-geraldo-do-baixio', 'sao-goncalo-do-abaete', 'sao-goncalo-do-para', 'sao-goncalo-do-rio-abaixo', 'sao-goncalo-do-sapucai', 'sao-gotardo', 'sao-joao-batista-do-gloria', 'sao-joao-da-lagoa', 'sao-joao-da-mata', 'sao-joao-da-ponte', 'sao-joao-das-missoes', 'sao-joao-del-rei', 'sao-joao-do-manhuacu', 'sao-joao-do-manteninha', 'sao-joao-do-oriente', 'sao-joao-do-pacui', 'sao-joao-do-paraiso', 'sao-joao-evangelista', 'sao-joao-nepomuceno', 'sao-joaquim-de-bicas', 'sao-jose-da-barra', 'sao-jose-da-lapa', 'sao-jose-da-safira', 'sao-jose-da-varginha', 'sao-jose-do-alegre', 'sao-jose-do-divino', 'sao-jose-do-goiabal', 'sao-jose-do-jacuri', 'sao-jose-do-mantimento', 'sao-lourenco', 'sao-miguel-do-anta', 'sao-pedro-da-uniao', 'sao-pedro-dos-ferros', 'sao-pedro-do-suacui', 'sao-romao', 'sao-roque-de-minas', 'sao-sebastiao-da-bela-vista', 'sao-sebastiao-da-vargem-alegre', 'sao-sebastiao-do-anta', 'sao-sebastiao-do-maranhao', 'sao-sebastiao-do-oeste', 'sao-sebastiao-do-paraiso', 'sao-sebastiao-do-rio-preto', 'sao-sebastiao-do-rio-verde', 'sao-tiago', 'sao-tomas-de-aquino', 'sao-tome-das-letras', 'sao-vicente-de-minas', 'sapucai-mirim', 'sardoa', 'sarzedo', 'setubinha', 'sem-peixe', 'senador-amaral', 'senador-cortes', 'senador-firmino', 'senador-jose-bento', 'senador-modestino-goncalves', 'senhora-de-oliveira', 'senhora-do-porto', 'senhora-dos-remedios', 'sericita', 'seritinga', 'serra-azul-de-minas', 'serra-da-saudade', 'serra-dos-aimores', 'serra-do-salitre', 'serrania', 'serranopolis-de-minas', 'serranos', 'sete-lagoas', 'silveirania', 'silvianopolis', 'simao-pereira', 'simonesia', 'sobralia', 'soledade-de-minas', 'tabuleiro', 'taiobeiras', 'taparuba', 'tapira', 'tapirai', 'taquaracu-de-minas', 'tarumirim', 'teixeiras', 'teofilo-otoni', 'timoteo', 'tiradentes', 'tiros', 'tocantins', 'tocos-do-moji', 'toledo', 'tombos', 'tres-coracoes', 'tres-marias', 'tres-pontas', 'tumiritinga', 'tupaciguara', 'turmalina', 'turvolandia', 'uba', 'ubai', 'ubaporanga', 'uberaba', 'uberlandia', 'umburatiba', 'unai', 'uniao-de-minas', 'uruana-de-minas', 'urucania', 'urucuia', 'vargem-alegre', 'vargem-bonita', 'vargem-grande-do-rio-pardo', 'varginha', 'varjao-de-minas', 'varzea-da-palma', 'varzelandia', 'vazante', 'verdelandia', 'veredinha', 'verissimo', 'vermelho-novo', 'vespasiano', 'vicosa', 'vieiras', 'mathias-lobato', 'virgem-da-lapa', 'virginia', 'virginopolis', 'virgolandia', 'visconde-do-rio-branco', 'volta-grande', 'wenceslau-braz'], 'MS': ['agua-clara', 'alcinopolis', 'amambai', 'anastacio', 'anaurilandia', 'angelica', 'antonio-joao', 'aparecida-do-taboado', 'aquidauana', 'aral-moreira', 'bandeirantes', 'bataguassu', 'bataypora', 'bela-vista', 'bodoquena', 'bonito', 'brasilandia', 'caarapo', 'camapua', 'campo-grande', 'caracol', 'cassilandia', 'chapadao-do-sul', 'corguinho', 'coronel-sapucaia', 'corumba', 'costa-rica', 'coxim', 'deodapolis', 'dois-irmaos-do-buriti', 'douradina', 'dourados', 'eldorado', 'fatima-do-sul', 'figueirao', 'gloria-de-dourados', 'guia-lopes-da-laguna', 'iguatemi', 'inocencia', 'itapora', 'itaquirai', 'ivinhema', 'japora', 'jaraguari', 'jardim', 'jatei', 'juti', 'ladario', 'laguna-carapa', 'maracaju', 'miranda', 'mundo-novo', 'navirai', 'nioaque', 'nova-alvorada-do-sul', 'nova-andradina', 'novo-horizonte-do-sul', 'paraiso-das-aguas', 'paranaiba', 'paranhos', 'pedro-gomes', 'ponta-pora', 'porto-murtinho', 'ribas-do-rio-pardo', 'rio-brilhante', 'rio-negro', 'rio-verde-de-mato-grosso', 'rochedo', 'santa-rita-do-pardo', 'sao-gabriel-do-oeste', 'sete-quedas', 'selviria', 'sidrolandia', 'sonora', 'tacuru', 'taquarussu', 'terenos', 'tres-lagoas', 'vicentina'], 'MT': ['santo-antonio-do-leverger', 'acorizal', 'agua-boa', 'alta-floresta', 'alto-araguaia', 'alto-boa-vista', 'alto-garcas', 'alto-paraguai', 'alto-taquari', 'apiacas', 'araguaiana', 'araguainha', 'araputanga', 'arenapolis', 'aripuana', 'barao-de-melgaco', 'barra-do-bugres', 'barra-do-garcas', 'bom-jesus-do-araguaia', 'brasnorte', 'caceres', 'campinapolis', 'campo-novo-do-parecis', 'campo-verde', 'campos-de-julio', 'canabrava-do-norte', 'canarana', 'carlinda', 'castanheira', 'chapada-dos-guimaraes', 'claudia', 'cocalinho', 'colider', 'colniza', 'comodoro', 'confresa', 'conquista-d-oeste', 'cotriguacu', 'cuiaba', 'curvelandia', 'denise', 'diamantino', 'dom-aquino', 'feliz-natal', 'figueiropolis-d-oeste', 'gaucha-do-norte', 'general-carneiro', 'gloria-d-oeste', 'guaranta-do-norte', 
-        'guiratinga', 'indiavai', 'ipiranga-do-norte', 'itanhanga', 'itauba', 'itiquira', 'jaciara', 'jangada', 'jauru', 'juara', 'juina', 'juruena', 'juscimeira', 'lambari-d-oeste', 'lucas-do-rio-verde', 'luciara', 'vila-bela-da-santissima-trindade', 'marcelandia', 'matupa', 'mirassol-d-oeste', 'nobres', 'nortelandia', 'nossa-senhora-do-livramento', 'nova-bandeirantes', 'nova-nazare', 'nova-lacerda', 'nova-santa-helena', 'nova-brasilandia', 'nova-canaa-do-norte', 'nova-mutum', 'nova-olimpia', 'nova-ubirata', 'nova-xavantina', 'novo-mundo', 'novo-horizonte-do-norte', 'novo-sao-joaquim', 'paranaita', 'paranatinga', 'novo-santo-antonio', 'pedra-preta', 'peixoto-de-azevedo', 'planalto-da-serra', 'pocone', 'pontal-do-araguaia', 'ponte-branca', 'pontes-e-lacerda', 'porto-alegre-do-norte', 'porto-dos-gauchos', 'porto-esperidiao', 'porto-estrela', 'poxoreo', 'primavera-do-leste', 'querencia', 'sao-jose-dos-quatro-marcos', 'reserva-do-cabacal', 'ribeirao-cascalheira', 'ribeiraozinho', 'rio-branco', 'santa-carmem', 'santo-afonso', 'sao-jose-do-povo', 'sao-jose-do-rio-claro', 'sao-jose-do-xingu', 'sao-pedro-da-cipa', 'rondolandia', 'rondonopolis', 'rosario-oeste', 'santa-cruz-do-xingu', 'salto-do-ceu', 'santa-rita-do-trivelato', 'santa-terezinha', 'santo-antonio-do-leste', 'santo-antonio-do-leverger', 'sao-felix-do-araguaia', 'sapezal', 'serra-nova-dourada', 'sinop', 'sorriso', 'tabapora', 'tangara-da-serra', 'tapurah', 'terra-nova-do-norte', 'tesouro', 'torixoreu', 'uniao-do-sul', 'vale-de-sao-domingos', 'varzea-grande', 'vera', 'vila-rica', 'nova-guarita', 'nova-marilandia', 'nova-maringa', 'nova-monte-verde', 'poxoreu'], 'PA': ['santa-izabel-do-para', 
-        'moju', 'abaetetuba', 'abel-figueiredo', 'acara', 'afua', 'agua-azul-do-norte', 'alenquer', 'almeirim', 'altamira', 'anajas', 'ananindeua', 'anapu', 'augusto-correa', 'aurora-do-para', 
-        'aveiro', 'bagre', 'baiao', 'bannach', 'barcarena', 'belem', 'belterra', 'benevides', 'bom-jesus-do-tocantins', 'bonito', 'braganca', 'brasil-novo', 'brejo-grande-do-araguaia', 'breu-branco', 'breves', 'bujaru', 'cachoeira-do-piria', 'cachoeira-do-arari', 'cameta', 'canaa-dos-carajas', 'capanema', 'capitao-poco', 'castanhal', 'chaves', 'colares', 'conceicao-do-araguaia', 'concordia-do-para', 'cumaru-do-norte', 'curionopolis', 'curralinho', 'curua', 'curuca', 'dom-eliseu', 'eldorado-dos-carajas', 'faro', 'floresta-do-araguaia', 'garrafao-do-norte', 'goianesia-do-para', 'gurupa', 'igarape-acu', 'igarape-miri', 'inhangapi', 'ipixuna-do-para', 'irituia', 'itaituba', 'itupiranga', 'jacareacanga', 'jacunda', 'juruti', 'limoeiro-do-ajuru', 'mae-do-rio', 'magalhaes-barata', 'maraba', 'maracana', 'marapanim', 'marituba', 'medicilandia', 'melgaco', 'mocajuba', 'mojui-dos-campos', 'monte-alegre', 'muana', 'nova-esperanca-do-piria', 'nova-ipixuna', 'nova-timboteua', 'novo-progresso', 'novo-repartimento', 'obidos', 'oeiras-do-para', 'oriximina', 'ourem', 'ourilandia-do-norte', 'pacaja', 'palestina-do-para', 'paragominas', 'parauapebas', 'pau-d-arco', 'peixe-boi', 'picarra', 'placas', 'ponta-de-pedras', 'portel', 'porto-de-moz', 'prainha', 'primavera', 'quatipuru', 'redencao', 'rio-maria', 'rondon-do-para', 'ruropolis', 'salinopolis', 'salvaterra', 'santa-barbara-do-para', 'santa-cruz-do-arari', 'santa-luzia-do-para', 'santa-maria-das-barreiras', 'santa-maria-do-para', 
-        'santana-do-araguaia', 'santarem', 'santarem-novo', 'santo-antonio-do-taua', 'sao-caetano-de-odivelas', 'sao-domingos-do-araguaia', 'sao-domingos-do-capim', 'sao-felix-do-xingu', 'sao-francisco-do-para', 'sao-geraldo-do-araguaia', 'sao-joao-da-ponta', 'sao-joao-de-pirabas', 'sao-joao-do-araguaia', 'sao-miguel-do-guama', 'sao-sebastiao-da-boa-vista', 'sapucaia', 'senador-jose-porfirio', 'soure', 'tailandia', 'terra-alta', 'terra-santa', 'tome-acu', 'tracuateua', 'trairao', 'tucuma', 'tucurui', 'ulianopolis', 'uruara', 'vigia', 'viseu', 'vitoria-do-xingu', 'xinguara'], 'PB': ['joca-claudino', 'campo-de-santana', 'agua-branca', 'aguiar', 'alagoa-grande', 'alagoa-nova', 'alagoinha', 'alcantil', 'algodao-de-jandaira', 'alhandra', 'sao-joao-do-rio-do-peixe', 'amparo', 'aparecida', 'aracagi', 'arara', 'araruna', 'areia', 'areia-de-baraunas', 'areial', 'aroeiras', 'assuncao', 'baia-da-traicao', 'bananeiras', 'barauna', 
-        'barra-de-santana', 'barra-de-santa-rosa', 'barra-de-sao-miguel', 'bayeux', 'belem', 'belem-do-brejo-do-cruz', 'bernardino-batista', 'boa-ventura', 'boa-vista', 'bom-jesus', 'bom-sucesso', 'bonito-de-santa-fe', 'boqueirao', 'igaracy', 'borborema', 'brejo-do-cruz', 'brejo-dos-santos', 'caapora', 'cabaceiras', 'cabedelo', 'cachoeira-dos-indios', 'cacimba-de-areia', 'cacimba-de-dentro', 'cacimbas', 'caicara', 'cajazeiras', 'cajazeirinhas', 'caldas-brandao', 'camalau', 'campina-grande', 'capim', 'caraubas', 'carrapateira', 'casserengue', 'catingueira', 
-        'catole-do-rocha', 'caturite', 'conceicao', 'condado', 'conde', 'congo', 'coremas', 'coxixola', 'cruz-do-espirito-santo', 'cubati', 'cuite', 'cuitegi', 'cuite-de-mamanguape', 'curral-de-cima', 'curral-velho', 'damiao', 'desterro', 'vista-serrana', 'diamante', 'dona-ines', 'duas-estradas', 'emas', 'esperanca', 'fagundes', 'frei-martinho', 'gado-bravo', 'guarabira', 'gurinhem', 'gurjao', 'ibiara', 'imaculada', 'inga', 'itabaiana', 'itaporanga', 'itapororoca', 'itatuba', 'jacarau', 'jerico', 'joao-pessoa', 'juarez-tavora', 'juazeirinho', 'junco-do-serido', 'juripiranga', 'juru', 'lagoa', 'lagoa-de-dentro', 'lagoa-seca', 'lastro', 'livramento', 'logradouro', 'lucena', 'mae-d-agua', 'malta', 'mamanguape', 'manaira', 'marcacao', 'mari', 'marizopolis', 'massaranduba', 'mataraca', 'matinhas', 'mato-grosso', 'matureia', 'mogeiro', 'montadas', 'monte-horebe', 'monteiro', 'mulungu', 'natuba', 'nazarezinho', 'nova-floresta', 'nova-olinda', 'nova-palmeira', 'olho-d-agua', 'olivedos', 'ouro-velho', 'parari', 'passagem', 'patos', 'paulista', 'pedra-branca', 'pedra-lavrada', 'pedras-de-fogo', 'pianco', 'picui', 'pilar', 'piloes', 'piloezinhos', 'pirpirituba', 'pitimbu', 'pocinhos', 'poco-dantas', 'poco-de-jose-de-moura', 'pombal', 'prata', 'princesa-isabel', 'puxinana', 'queimadas', 'quixaba', 'remigio', 'pedro-regis', 'riachao', 'riachao-do-bacamarte', 'riachao-do-poco', 'riacho-de-santo-antonio', 'riacho-dos-cavalos', 'rio-tinto', 'salgadinho', 'salgado-de-sao-felix', 'santa-cecilia', 'santa-cruz', 'santa-helena', 'santa-ines', 'santa-luzia', 'santana-de-mangueira', 'santana-dos-garrotes', 'santarem', 'santa-rita', 'santa-teresinha', 'santo-andre', 'sao-bento', 'sao-bentinho', 'sao-domingos-do-cariri', 'sao-domingos', 'sao-francisco', 'sao-joao-do-cariri', 'sao-joao-do-tigre', 'sao-jose-da-lagoa-tapada', 'sao-jose-de-caiana', 'sao-jose-de-espinharas', 'sao-jose-dos-ramos', 'sao-jose-de-piranhas', 'sao-jose-de-princesa', 'sao-jose-do-bonfim', 'sao-jose-do-brejo-do-cruz', 'sao-jose-do-sabugi', 'sao-jose-dos-cordeiros', 'sao-mamede', 'sao-miguel-de-taipu', 'sao-sebastiao-de-lagoa-de-roca', 'sao-sebastiao-do-umbuzeiro', 'sape', 'serido', 'serra-branca', 'serra-da-raiz', 'serra-grande', 'serra-redonda', 'serraria', 'sertaozinho', 'sobrado', 'solanea', 'soledade', 'sossego', 'sousa', 'sume', 'tacima', 'taperoa', 'tavares', 'teixeira', 'tenorio', 'triunfo', 'uirauna', 'umbuzeiro', 'varzea', 'vieiropolis', 'zabele', 'sao-vicente-do-serido'], 'PE': ['lagoa-de-itaenga', 'abreu-e-lima', 'afogados-da-ingazeira', 'afranio', 'agrestina', 'agua-preta', 'aguas-belas', 'alagoinha', 'alianca', 'altinho', 'amaraji', 'angelim', 'aracoiaba', 'araripina', 'arcoverde', 'barra-de-guabiraba', 'barreiros', 'belem-de-maria', 'belem-do-sao-francisco', 'belo-jardim', 
-        'betania', 'bezerros', 'bodoco', 'bom-conselho', 'bom-jardim', 'bonito', 'brejao', 'brejinho', 'brejo-da-madre-de-deus', 'buenos-aires', 'buique', 'cabo-de-santo-agostinho', 'cabrobo', 
-        'cachoeirinha', 'caetes', 'calcado', 'calumbi', 'camaragibe', 'camocim-de-sao-felix', 'camutanga', 'canhotinho', 'capoeiras', 'carnaiba', 'carnaubeira-da-penha', 'carpina', 'caruaru', 'casinhas', 'catende', 'cedro', 'cha-de-alegria', 'cha-grande', 'condado', 'correntes', 'cortes', 'cumaru', 'cupira', 'custodia', 'dormentes', 'escada', 'exu', 'feira-nova', 'fernando-de-noronha', 'ferreiros', 'flores', 'floresta', 'frei-miguelinho', 'gameleira', 'garanhuns', 'gloria-do-goita', 'goiana', 'granito', 'gravata', 'iati', 'ibimirim', 'ibirajuba', 'igarassu', 'iguaraci', 'inaja', 'ingazeira', 'ipojuca', 'ipubi', 'itacuruba', 'itaiba', 'ilha-de-itamaraca', 'itambe', 'itapetim', 'itapissuma', 'itaquitinga', 'jaboatao-dos-guararapes', 'jaqueira', 'jatauba', 'jatoba', 'joao-alfredo', 'joaquim-nabuco', 'jucati', 'jupi', 'jurema', 'lagoa-do-carro', 'lagoa-do-itaenga', 'lagoa-do-ouro', 'lagoa-dos-gatos', 'lagoa-grande', 'lajedo', 'limoeiro', 'macaparana', 'machados', 'manari', 'maraial', 'mirandiba', 'moreno', 'nazare-da-mata', 'olinda', 'orobo', 'oroco', 'ouricuri', 'palmares', 'palmeirina', 'panelas', 'paranatama', 'parnamirim', 'passira', 'paudalho', 'paulista', 'pedra', 'pesqueira', 'petrolandia', 'petrolina', 'pocao', 'pombos', 'primavera', 'quipapa', 'quixaba', 'recife', 'riacho-das-almas', 'ribeirao', 'rio-formoso', 'saire', 'salgadinho', 'salgueiro', 'saloa', 'sanharo', 'santa-cruz', 'santa-cruz-da-baixa-verde', 'santa-cruz-do-capibaribe', 'santa-filomena', 'santa-maria-da-boa-vista', 'santa-maria-do-cambuca', 'santa-terezinha', 'sao-benedito-do-sul', 'sao-bento-do-una', 'sao-caitano', 'sao-joao', 'sao-joaquim-do-monte', 'sao-jose-da-coroa-grande', 'sao-jose-do-belmonte', 'sao-jose-do-egito', 'sao-lourenco-da-mata', 'sao-vicente-ferrer', 'serra-talhada', 'serrita', 'sertania', 'sirinhaem', 'moreilandia', 'solidao', 'surubim', 
-        'tabira', 'tacaimbo', 'tacaratu', 'tamandare', 'taquaritinga-do-norte', 'terezinha', 'terra-nova', 'timbauba', 'toritama', 'tracunhaem', 'trindade', 'triunfo', 'tupanatinga', 'tuparetama', 'venturosa', 'verdejante', 'vertente-do-lerio', 'vertentes', 'vicencia', 'vitoria-de-santo-antao', 'xexeu'], 'PI': ['batalha', 'acaua', 'agricolandia', 'agua-branca', 'alagoinha-do-piaui', 'alegrete-do-piaui', 'alto-longa', 'altos', 'alvorada-do-gurgueia', 'amarante', 'angical-do-piaui', 'anisio-de-abreu', 'antonio-almeida', 'aroazes', 'aroeiras-do-itaim', 'arraial', 'assuncao-do-piaui', 'avelino-lopes', 'baixa-grande-do-ribeiro', 'barra-d-alcantara', 'barras', 'barreiras-do-piaui', 'barro-duro', 'bela-vista-do-piaui', 'belem-do-piaui', 'beneditinos', 'bertolinia', 'betania-do-piaui', 'boa-hora', 'bocaina', 'bom-jesus', 'bom-principio-do-piaui', 'bonfim-do-piaui', 'boqueirao-do-piaui', 'brasileira', 'brejo-do-piaui', 'buriti-dos-lopes', 'buriti-dos-montes', 'cabeceiras-do-piaui', 'cajazeiras-do-piaui', 'cajueiro-da-praia', 'caldeirao-grande-do-piaui', 'campinas-do-piaui', 'campo-alegre-do-fidalgo', 'campo-grande-do-piaui', 'campo-largo-do-piaui', 'campo-maior', 'canavieira', 'canto-do-buriti', 'capitao-de-campos', 'capitao-gervasio-oliveira', 'caracol', 'caraubas-do-piaui', 'caridade-do-piaui', 'castelo-do-piaui', 'caxingo', 'cocal', 'cocal-de-telha', 'cocal-dos-alves', 'coivaras', 'colonia-do-gurgueia', 'colonia-do-piaui', 'conceicao-do-caninde', 'coronel-jose-dias', 'corrente', 'cristalandia-do-piaui', 'cristino-castro', 'curimata', 'currais', 'curralinhos', 'curral-novo-do-piaui', 'demerval-lobao', 'dirceu-arcoverde', 'dom-expedito-lopes', 'domingos-mourao', 'dom-inocencio', 'elesbao-veloso', 'eliseu-martins', 'esperantina', 'fartura-do-piaui', 'flores-do-piaui', 'floresta-do-piaui', 'floriano', 'francinopolis', 'francisco-ayres', 'francisco-macedo', 'francisco-santos', 'fronteiras', 'geminiano', 'gilbues', 'guadalupe', 'guaribas', 'hugo-napoleao', 'ilha-grande', 'inhuma', 'ipiranga-do-piaui', 'isaias-coelho', 'itainopolis', 'itaueira', 'jacobina-do-piaui', 'jaicos', 'jardim-do-mulato', 'jatoba-do-piaui', 'jerumenha', 'joao-costa', 'joaquim-pires', 'joca-marques', 'jose-de-freitas', 'juazeiro-do-piaui', 'julio-borges', 'jurema', 'lagoinha-do-piaui', 'lagoa-alegre', 'lagoa-do-barro-do-piaui', 'lagoa-de-sao-francisco', 'lagoa-do-piaui', 'lagoa-do-sitio', 'landri-sales', 'luis-correia', 'luzilandia', 'madeiro', 'manoel-emidio', 'marcolandia', 'marcos-parente', 'massape-do-piaui', 'matias-olimpio', 'miguel-alves', 'miguel-leao', 'milton-brandao', 'monsenhor-gil', 'monsenhor-hipolito', 'monte-alegre-do-piaui', 'morro-cabeca-no-tempo', 'morro-do-chapeu-do-piaui', 'murici-dos-portelas', 'nazare-do-piaui', 'nazaria', 'nossa-senhora-de-nazare', 'nossa-senhora-dos-remedios', 'novo-oriente-do-piaui', 'novo-santo-antonio', 'oeiras', 'olho-d-agua-do-piaui', 'padre-marcos', 'paes-landim', 'pajeu-do-piaui', 'palmeira-do-piaui', 'palmeirais', 'paqueta', 'parnagua', 'parnaiba', 'passagem-franca-do-piaui', 'patos-do-piaui', 'pau-d-arco-do-piaui', 'paulistana', 'pavussu', 'pedro-ii', 'pedro-laurentino', 'nova-santa-rita', 'picos', 'pimenteiras', 'pio-ix', 'piracuruca', 'piripiri', 'porto', 'porto-alegre-do-piaui', 'prata-do-piaui', 'queimada-nova', 'redencao-do-gurgueia', 'regeneracao', 'riacho-frio', 'ribeira-do-piaui', 'ribeiro-goncalves', 'rio-grande-do-piaui', 'santa-cruz-do-piaui', 'santa-cruz-dos-milagres', 'santa-filomena', 'santa-luz', 'santana-do-piaui', 'santa-rosa-do-piaui', 'santo-antonio-de-lisboa', 'santo-antonio-dos-milagres', 'santo-inacio-do-piaui', 'sao-braz-do-piaui', 'sao-felix-do-piaui', 'sao-francisco-de-assis-do-piaui', 'sao-francisco-do-piaui', 'sao-goncalo-do-gurgueia', 'sao-goncalo-do-piaui', 'sao-joao-da-canabrava', 'sao-joao-da-fronteira', 'sao-joao-da-serra', 'sao-joao-da-varjota', 'sao-joao-do-arraial', 'sao-joao-do-piaui', 'sao-jose-do-divino', 'sao-jose-do-peixe', 'sao-jose-do-piaui', 'sao-juliao', 'sao-lourenco-do-piaui', 'sao-luis-do-piaui', 'sao-miguel-da-baixa-grande', 'sao-miguel-do-fidalgo', 'sao-miguel-do-tapuio', 'sao-pedro-do-piaui', 'sao-raimundo-nonato', 'sebastiao-barros', 'sebastiao-leal', 'sigefredo-pacheco', 'simoes', 'simplicio-mendes', 'socorro-do-piaui', 'sussuapara', 'tamboril-do-piaui', 'tanque-do-piaui', 'teresina', 'uniao', 'urucui', 'valenca-do-piaui', 'varzea-branca', 'varzea-grande', 'vera-mendes', 'vila-nova-do-piaui', 'wall-ferraz'], 'PR': ['abatia', 'adrianopolis', 'agudos-do-sul', 'almirante-tamandare', 'altamira-do-parana', 'altonia', 'alto-parana', 'alto-piquiri', 'alvorada-do-sul', 'amapora', 'ampere', 'anahy', 'andira', 'angulo', 'antonina', 'antonio-olinto', 'apucarana', 'arapongas', 'arapoti', 'arapua', 'araruna', 'araucaria', 'ariranha-do-ivai', 'assai', 'assis-chateaubriand', 'astorga', 'atalaia', 'balsa-nova', 'bandeirantes', 'barbosa-ferraz', 'barracao', 'barra-do-jacare', 'bela-vista-da-caroba', 'bela-vista-do-paraiso', 'bituruna', 'boa-esperanca', 'boa-esperanca-do-iguacu', 'boa-ventura-de-sao-roque', 'boa-vista-da-aparecida', 'bocaiuva-do-sul', 'bom-jesus-do-sul', 'bom-sucesso', 'bom-sucesso-do-sul', 'borrazopolis', 'braganey', 'brasilandia-do-sul', 'cafeara', 'cafelandia', 'cafezal-do-sul', 'california', 'cambara', 'cambe', 'cambira', 'campina-da-lagoa', 'campina-do-simao', 'campina-grande-do-sul', 'campo-bonito', 'campo-do-tenente', 'campo-largo', 'campo-magro', 'campo-mourao', 'candido-de-abreu', 'candoi', 'cantagalo', 'capanema', 'capitao-leonidas-marques', 'carambei', 'carlopolis', 'cascavel', 'castro', 'catanduvas', 'centenario-do-sul', 'cerro-azul', 'ceu-azul', 'chopinzinho', 'cianorte', 'cidade-gaucha', 'clevelandia', 'colombo', 'colorado', 'congonhinhas', 'conselheiro-mairinck', 'contenda', 'corbelia', 'cornelio-procopio', 'coronel-domingos-soares', 'coronel-vivida', 'corumbatai-do-sul', 'cruzeiro-do-iguacu', 'cruzeiro-do-oeste', 'cruzeiro-do-sul', 'cruz-machado', 'cruzmaltina', 'curitiba', 'curiuva', 'diamante-do-norte', 'diamante-do-sul', 'diamante-d-oeste', 'dois-vizinhos', 'douradina', 'doutor-camargo', 'eneas-marques', 'engenheiro-beltrao', 'esperanca-nova', 'entre-rios-do-oeste', 'espigao-alto-do-iguacu', 'farol', 'faxinal', 'fazenda-rio-grande', 'fenix', 'fernandes-pinheiro', 'figueira', 'florai', 'flor-da-serra-do-sul', 'floresta', 'florestopolis', 'florida', 'formosa-do-oeste', 'foz-do-iguacu', 'francisco-alves', 'francisco-beltrao', 'foz-do-jordao', 'general-carneiro', 'godoy-moreira', 'goioere', 'goioxim', 'grandes-rios', 'guaira', 'guairaca', 'guamiranga', 'guapirama', 'guaporema', 'guaraci', 'guaraniacu', 'guarapuava', 'guaraquecaba', 'guaratuba', 'honorio-serpa', 'ibaiti', 'ibema', 'ibipora', 'icaraima', 'iguaracu', 'iguatu', 'imbau', 'imbituva', 'inacio-martins', 'inaja', 'indianopolis', 'ipiranga', 'ipora', 'iracema-do-oeste', 'irati', 'iretama', 'itaguaje', 'itaipulandia', 'itambaraca', 'itambe', 'itapejara-d-oeste', 'itaperucu', 'itauna-do-sul', 'ivai', 'ivaipora', 'ivate', 'ivatuba', 'jaboti', 'jacarezinho', 'jaguapita', 'jaguariaiva', 'jandaia-do-sul', 'janiopolis', 'japira', 'japura', 'jardim-alegre', 'jardim-olinda', 'jataizinho', 'jesuitas', 'joaquim-tavora', 'jundiai-do-sul', 'juranda', 'jussara', 'kalore', 'lapa', 'laranjal', 'laranjeiras-do-sul', 'leopolis', 'lidianopolis', 'lindoeste', 'loanda', 'lobato', 'londrina', 'luiziana', 'lunardelli', 'lupionopolis', 'mallet', 'mambore', 'mandaguacu', 'mandaguari', 'mandirituba', 'manfrinopolis', 'mangueirinha', 'manoel-ribas', 'marechal-candido-rondon', 'maria-helena', 'marialva', 'marilandia-do-sul', 'marilena', 'mariluz', 'maringa', 'mariopolis', 'maripa', 'marmeleiro', 'marquinho', 'marumbi', 'matelandia', 'matinhos', 'mato-rico', 'maua-da-serra', 'medianeira', 'mercedes', 'mirador', 'miraselva', 'missal', 'moreira-sales', 'morretes', 'munhoz-de-melo', 'nossa-senhora-das-gracas', 'nova-alianca-do-ivai', 'nova-america-da-colina', 'nova-aurora', 'nova-cantu', 'nova-esperanca', 'nova-esperanca-do-sudoeste', 'nova-fatima', 'nova-laranjeiras', 'nova-londrina', 'nova-olimpia', 'nova-santa-barbara', 'nova-santa-rosa', 'nova-prata-do-iguacu', 'nova-tebas', 'novo-itacolomi', 'ortigueira', 'ourizona', 'ouro-verde-do-oeste', 'paicandu', 'palmas', 'palmeira', 'palmital', 'palotina', 'paraiso-do-norte', 'paranacity', 'paranagua', 'paranapoema', 'paranavai', 'pato-bragado', 'pato-branco', 'paula-freitas', 'paulo-frontin', 'peabiru', 'perobal', 'perola', 'perola-d-oeste', 'pien', 'pinhais', 'pinhalao', 'pinhal-de-sao-bento', 'pinhao', 'pirai-do-sul', 'piraquara', 'pitanga', 'pitangueiras', 'planaltina-do-parana', 'planalto', 'ponta-grossa', 'pontal-do-parana', 'porecatu', 'porto-amazonas', 'porto-barreiro', 'porto-rico', 
-        'porto-vitoria', 'prado-ferreira', 'pranchita', 'presidente-castelo-branco', 'primeiro-de-maio', 'prudentopolis', 'quarto-centenario', 'quatigua', 'quatro-barras', 'quatro-pontes', 'quedas-do-iguacu', 'querencia-do-norte', 'quinta-do-sol', 'quitandinha', 'ramilandia', 'rancho-alegre', 'rancho-alegre-d-oeste', 'realeza', 'reboucas', 'renascenca', 'reserva', 'reserva-do-iguacu', 'ribeirao-claro', 'ribeirao-do-pinhal', 'rio-azul', 'rio-bom', 'rio-bonito-do-iguacu', 'rio-branco-do-ivai', 'rio-branco-do-sul', 'rio-negro', 'rolandia', 'roncador', 'rondon', 'rosario-do-ivai', 'sabaudia', 'salgado-filho', 'salto-do-itarare', 'salto-do-lontra', 'santa-amelia', 'santa-cecilia-do-pavao', 'santa-cruz-de-monte-castelo', 'santa-fe', 'santa-helena', 'santa-ines', 'santa-isabel-do-ivai', 'santa-izabel-do-oeste', 'santa-lucia', 'santa-maria-do-oeste', 'santa-mariana', 'santa-monica', 'santana-do-itarare', 'santa-tereza-do-oeste', 'santa-terezinha-de-itaipu', 'santo-antonio-da-platina', 'santo-antonio-do-caiua', 'santo-antonio-do-paraiso', 'santo-antonio-do-sudoeste', 'santo-inacio', 'sao-carlos-do-ivai', 'sao-jeronimo-da-serra', 'sao-joao', 'sao-joao-do-caiua', 'sao-joao-do-ivai', 'sao-joao-do-triunfo', 'sao-jorge-d-oeste', 'sao-jorge-do-ivai', 'sao-jorge-do-patrocinio', 'sao-jose-da-boa-vista', 'sao-jose-das-palmeiras', 'sao-jose-dos-pinhais', 'sao-manoel-do-parana', 'sao-mateus-do-sul', 'sao-miguel-do-iguacu', 'sao-pedro-do-iguacu', 'sao-pedro-do-ivai', 'sao-pedro-do-parana', 'sao-sebastiao-da-amoreira', 'sao-tome', 'sapopema', 'sarandi', 'saudade-do-iguacu', 'senges', 'serranopolis-do-iguacu', 'sertaneja', 'sertanopolis', 'siqueira-campos', 'sulina', 
-        'tamarana', 'tamboara', 'tapejara', 'tapira', 'teixeira-soares', 'telemaco-borba', 'terra-boa', 'terra-rica', 'terra-roxa', 'tibagi', 'tijucas-do-sul', 'toledo', 'tomazina', 'tres-barras-do-parana', 'tunas-do-parana', 'tuneiras-do-oeste', 'tupassi', 'turvo', 'ubirata', 'umuarama', 'uniao-da-vitoria', 'uniflor', 'urai', 'wenceslau-braz', 'ventania', 'vera-cruz-do-oeste', 'vere', 'alto-paraiso', 'doutor-ulysses', 'virmond', 'vitorino', 'xambre'], 'RJ': ['angra-dos-reis', 'aperibe', 'araruama', 'areal', 'armacao-de-buzios', 'arraial-do-cabo', 'barra-do-pirai', 'barra-mansa', 'belford-roxo', 'bom-jardim', 'bom-jesus-do-itabapoana', 'cabo-frio', 'cachoeiras-de-macacu', 'cambuci', 'carapebus', 'comendador-levy-gasparian', 'campos-dos-goytacazes', 'cantagalo', 'cardoso-moreira', 'carmo', 'casimiro-de-abreu', 'conceicao-de-macabu', 'cordeiro', 'duas-barras', 'duque-de-caxias', 'engenheiro-paulo-de-frontin', 'guapimirim', 'iguaba-grande', 'itaborai', 'itaguai', 'italva', 'itaocara', 'itaperuna', 'itatiaia', 'japeri', 'laje-do-muriae', 'macae', 'macuco', 'mage', 'mangaratiba', 'marica', 'mendes', 'mesquita', 'miguel-pereira', 'miracema', 'natividade', 'nilopolis', 'niteroi', 'nova-friburgo', 'nova-iguacu', 'paracambi', 'paraiba-do-sul', 'paraty', 'paty-do-alferes', 'petropolis', 'pinheiral', 'pirai', 'porciuncula', 'porto-real', 'quatis', 'queimados', 'quissama', 'resende', 'rio-bonito', 'rio-claro', 'rio-das-flores', 'rio-das-ostras', 'rio-de-janeiro', 'santa-maria-madalena', 'santo-antonio-de-padua', 'sao-francisco-de-itabapoana', 'sao-fidelis', 'sao-goncalo', 'sao-joao-da-barra', 'sao-joao-de-meriti', 'sao-jose-de-uba', 'sao-jose-do-vale-do-rio-preto', 'sao-pedro-da-aldeia', 'sao-sebastiao-do-alto', 'sapucaia', 'saquarema', 'seropedica', 'silva-jardim', 'sumidouro', 'tangua', 'teresopolis', 'trajano-de-moraes', 'tres-rios', 'valenca', 'varre-sai', 'vassouras', 'volta-redonda'], 'RN': ['serra-caiada', 'acari', 'acu', 'afonso-bezerra', 'agua-nova', 'alexandria', 'almino-afonso', 'alto-do-rodrigues', 'angicos', 'antonio-martins', 'apodi', 'areia-branca', 'ares', 'augusto-severo', 'baia-formosa', 'barauna', 'barcelona', 'bento-fernandes', 'bodo', 'bom-jesus', 'brejinho', 'caicara-do-norte', 'caicara-do-rio-do-vento', 'caico', 'campo-redondo', 'canguaretama', 'caraubas', 'carnauba-dos-dantas', 'carnaubais', 'ceara-mirim', 'cerro-cora', 'coronel-ezequiel', 'coronel-joao-pessoa', 'cruzeta', 'currais-novos', 'doutor-severiano', 'parnamirim', 'encanto', 'equador', 'espirito-santo', 'extremoz', 'felipe-guerra', 'fernando-pedroza', 'florania', 'francisco-dantas', 'frutuoso-gomes', 'galinhos', 'goianinha', 'governador-dix-sept-rosado', 'grossos', 'guamare', 'ielmo-marinho', 'ipanguacu', 'ipueira', 'itaja', 'itau', 'jacana', 'jandaira', 'janduis', 'januario-cicco', 'japi', 'jardim-de-angicos', 'jardim-de-piranhas', 'jardim-do-serido', 'joao-camara', 'joao-dias', 'jose-da-penha', 'jucurutu', 'jundia', 'lagoa-d-anta', 'lagoa-de-pedras', 'lagoa-de-velhos', 'lagoa-nova', 'lagoa-salgada', 'lajes', 'lajes-pintadas', 'lucrecia', 'luis-gomes', 'macaiba', 'macau', 'major-sales', 'marcelino-vieira', 'martins', 'maxaranguape', 'messias-targino', 'montanhas', 'monte-alegre', 'monte-das-gameleiras', 'mossoro', 'natal', 'nisia-floresta', 'nova-cruz', 'olho-d-agua-do-borges', 'ouro-branco', 'parana', 'parau', 'parazinho', 'parelhas', 'rio-do-fogo', 'passa-e-fica', 'passagem', 'patu', 'santa-maria', 'pau-dos-ferros', 'pedra-grande', 'pedra-preta', 'pedro-avelino', 'pedro-velho', 'pendencias', 'piloes', 'poco-branco', 'portalegre', 'porto-do-mangue', 'presidente-juscelino', 'pureza', 'rafael-fernandes', 'rafael-godeiro', 'riacho-da-cruz', 'riacho-de-santana', 'riachuelo', 'rodolfo-fernandes', 'tibau', 'ruy-barbosa', 'santa-cruz', 'santana-do-matos', 'santana-do-serido', 'santo-antonio', 'sao-bento-do-norte', 'sao-bento-do-trairi', 'sao-fernando', 'sao-francisco-do-oeste', 'sao-goncalo-do-amarante', 'sao-joao-do-sabugi', 'sao-jose-de-mipibu', 'sao-jose-do-campestre', 'sao-jose-do-serido', 'sao-miguel', 'sao-miguel-do-gostoso', 'sao-paulo-do-potengi', 'sao-pedro', 'sao-rafael', 'sao-tome', 'sao-vicente', 'senador-eloi-de-souza', 'senador-georgino-avelino', 'serra-de-sao-bento', 'serra-do-mel', 'serra-negra-do-norte', 'serrinha', 'serrinha-dos-pintos', 'severiano-melo', 'sitio-novo', 'taboleiro-grande', 'taipu', 'tangara', 'tenente-ananias', 'tenente-laurentino-cruz', 'tibau-do-sul', 'timbauba-dos-batistas', 'touros', 'triunfo-potiguar', 'umarizal', 'upanema', 'varzea', 'venha-ver', 'vera-cruz', 'vicosa', 'vila-flor'], 'RO': ['alta-floresta-d-oeste', 'ariquemes', 'cabixi', 'cacoal', 'cerejeiras', 'colorado-do-oeste', 'corumbiara', 'costa-marques', 'espigao-d-oeste', 'guajara-mirim', 'jaru', 'ji-parana', 'machadinho-d-oeste', 'nova-brasilandia-d-oeste', 'ouro-preto-do-oeste', 'pimenta-bueno', 'porto-velho', 'presidente-medici', 'rio-crespo', 'rolim-de-moura', 'santa-luzia-d-oeste', 'vilhena', 'sao-miguel-do-guapore', 'nova-mamore', 'alvorada-d-oeste', 'alto-alegre-dos-parecis', 'alto-paraiso', 'buritis', 'novo-horizonte-do-oeste', 'cacaulandia', 'campo-novo-de-rondonia', 'candeias-do-jamari', 'castanheiras', 'chupinguaia', 
-        'cujubim', 'governador-jorge-teixeira', 'itapua-do-oeste', 'ministro-andreazza', 'mirante-da-serra', 'monte-negro', 'nova-uniao', 'parecis', 'pimenteiras-do-oeste', 'primavera-de-rondonia', 'sao-felipe-d-oeste', 'sao-francisco-do-guapore', 'seringueiras', 'teixeiropolis', 'theobroma', 'urupa', 'vale-do-anari', 'vale-do-paraiso'], 'RR': ['amajari', 'alto-alegre', 'boa-vista', 'bonfim', 'canta', 'caracarai', 'caroebe', 'iracema', 'mucajai', 'normandia', 'pacaraima', 'rorainopolis', 'sao-joao-da-baliza', 'sao-luiz', 'uiramuta'], 'RS': ['bossoroca', 'acegua', 'agua-santa', 'agudo', 'ajuricaba', 'alecrim', 'alegrete', 'alegria', 'almirante-tamandare-do-sul', 'alpestre', 'alto-alegre', 'alto-feliz', 'alvorada', 'amaral-ferrador', 'ametista-do-sul', 'andre-da-rocha', 'anta-gorda', 'antonio-prado', 'arambare', 'ararica', 'aratiba', 'arroio-do-meio', 'arroio-do-sal', 'arroio-do-padre', 'arroio-dos-ratos', 'arroio-do-tigre', 'arroio-grande', 'arvorezinha', 'augusto-pestana', 'aurea', 'bage', 'balneario-pinhal', 'barao', 'barao-de-cotegipe', 'barao-do-triunfo', 'barracao', 'barra-do-guarita', 'barra-do-quarai', 'barra-do-ribeiro', 'barra-do-rio-azul', 'barra-funda', 'barros-cassal', 'benjamin-constant-do-sul', 'bento-goncalves', 'boa-vista-das-missoes', 'boa-vista-do-burica', 'boa-vista-do-cadeado', 'boa-vista-do-incra', 'boa-vista-do-sul', 'bom-jesus', 'bom-principio', 'bom-progresso', 'bom-retiro-do-sul', 'boqueirao-do-leao', 'bozano', 'braga', 'brochier', 'butia', 'cacapava-do-sul', 'cacequi', 'cachoeira-do-sul', 'cachoeirinha', 'cacique-doble', 'caibate', 'caicara', 'camaqua', 'camargo', 'cambara-do-sul', 'campestre-da-serra', 'campina-das-missoes', 'campinas-do-sul', 'campo-bom', 'campo-novo', 'campos-borges', 'candelaria', 'candido-godoi', 'candiota', 'canela', 'cangucu', 'canoas', 'canudos-do-vale', 'capao-bonito-do-sul', 
-        'capao-da-canoa', 'capao-do-cipo', 'capao-do-leao', 'capivari-do-sul', 'capela-de-santana', 'capitao', 'carazinho', 'caraa', 'carlos-barbosa', 'carlos-gomes', 'casca', 'caseiros', 'catuipe', 'caxias-do-sul', 'centenario', 'cerrito', 'cerro-branco', 'cerro-grande', 'cerro-grande-do-sul', 'cerro-largo', 'chapada', 'charqueadas', 'charrua', 'chiapetta', 'chui', 'chuvisca', 'cidreira', 'ciriaco', 'colinas', 'colorado', 'condor', 'constantina', 'coqueiro-baixo', 'coqueiros-do-sul', 'coronel-barros', 'coronel-bicaco', 'coronel-pilar', 'cotipora', 'coxilha', 'crissiumal', 'cristal', 'cristal-do-sul', 'cruz-alta', 'cruzaltense', 'cruzeiro-do-sul', 'david-canabarro', 'derrubadas', 'dezesseis-de-novembro', 'dilermando-de-aguiar', 'dois-irmaos', 'dois-irmaos-das-missoes', 'dois-lajeados', 'dom-feliciano', 'dom-pedro-de-alcantara', 'dom-pedrito', 'dona-francisca', 'doutor-mauricio-cardoso', 'doutor-ricardo', 'eldorado-do-sul', 'encantado', 'encruzilhada-do-sul', 'engenho-velho', 'entre-ijuis', 'entre-rios-do-sul', 'erebango', 'erechim', 'ernestina', 'herval', 'erval-grande', 'erval-seco', 'esmeralda', 'esperanca-do-sul', 'espumoso', 'estacao', 'estancia-velha', 'esteio', 'estrela', 'estrela-velha', 'eugenio-de-castro', 'fagundes-varela', 'farroupilha', 'faxinal-do-soturno', 'faxinalzinho', 'fazenda-vilanova', 'feliz', 'flores-da-cunha', 'floriano-peixoto', 'fontoura-xavier', 'formigueiro', 'forquetinha', 'fortaleza-dos-valos', 'frederico-westphalen', 'garibaldi', 'garruchos', 'gaurama', 'general-camara', 'gentil', 'getulio-vargas', 'girua', 'glorinha', 'gramado', 'gramado-dos-loureiros', 'gramado-xavier', 'gravatai', 'guabiju', 'guaiba', 'guapore', 
-        'guarani-das-missoes', 'harmonia', 'herveiras', 'horizontina', 'hulha-negra', 'humaita', 'ibarama', 'ibiaca', 'ibiraiaras', 'ibirapuita', 'ibiruba', 'igrejinha', 'ijui', 'ilopolis', 'imbe', 'imigrante', 'independencia', 'inhacora', 'ipe', 'ipiranga-do-sul', 'irai', 'itaara', 'itacurubi', 'itapuca', 'itaqui', 'itati', 'itatiba-do-sul', 'ivora', 'ivoti', 'jaboticaba', 'jacuizinho', 'jacutinga', 'jaguarao', 'jaguari', 'jaquirana', 'jari', 'joia', 'julio-de-castilhos', 'lagoa-bonita-do-sul', 'lagoao', 'lagoa-dos-tres-cantos', 'lagoa-vermelha', 'lajeado', 'lajeado-do-bugre', 'lavras-do-sul', 'liberato-salzano', 'lindolfo-collor', 'linha-nova', 'machadinho', 'macambara', 'mampituba', 'manoel-viana', 'maquine', 'marata', 'marau', 'marcelino-ramos', 'mariana-pimentel', 'mariano-moro', 'marques-de-souza', 'mata', 'mato-castelhano', 'mato-leitao', 'mato-queimado', 'maximiliano-de-almeida', 'minas-do-leao', 'miraguai', 'montauri', 'monte-alegre-dos-campos', 'monte-belo-do-sul', 'montenegro', 'mormaco', 'morrinhos-do-sul', 'morro-redondo', 'morro-reuter', 'mostardas', 'mucum', 'muitos-capoes', 'muliterno', 'nao-me-toque', 'nicolau-vergueiro', 'nonoai', 'nova-alvorada', 'nova-araca', 'nova-bassano', 'nova-boa-vista', 'nova-brescia', 'nova-candelaria', 'nova-esperanca-do-sul', 'nova-hartz', 'nova-padua', 'nova-palma', 'nova-petropolis', 'nova-prata', 'nova-ramada', 'nova-roma-do-sul', 'nova-santa-rita', 'novo-cabrais', 'novo-hamburgo', 'novo-machado', 'novo-tiradentes', 'novo-xingu', 'novo-barreiro', 'osorio', 'paim-filho', 'palmares-do-sul', 'palmeira-das-missoes', 'palmitinho', 'panambi', 'pantano-grande', 'parai', 'paraiso-do-sul', 'pareci-novo', 'parobe', 'passa-sete', 'passo-do-sobrado', 'passo-fundo', 'paulo-bento', 'paverama', 'pedras-altas', 'pedro-osorio', 'pejucara', 'pelotas', 'picada-cafe', 'pinhal', 'pinhal-da-serra', 'pinhal-grande', 'pinheirinho-do-vale', 'pinheiro-machado', 'pinto-bandeira', 'pirapo', 'piratini', 'planalto', 'poco-das-antas', 'pontao', 'ponte-preta', 'portao', 'porto-alegre', 'porto-lucena', 'porto-maua', 'porto-vera-cruz', 'porto-xavier', 'pouso-novo', 'presidente-lucena', 'progresso', 'protasio-alves', 'putinga', 'quarai', 'quatro-irmaos', 'quevedos', 'quinze-de-novembro', 'redentora', 'relvado', 'restinga-seca', 'rio-dos-indios', 'rio-grande', 'rio-pardo', 'riozinho', 'roca-sales', 'rodeio-bonito', 'rolador', 'rolante', 'ronda-alta', 'rondinha', 'roque-gonzales', 'rosario-do-sul', 'sagrada-familia', 'saldanha-marinho', 'salto-do-jacui', 'salvador-das-missoes', 'salvador-do-sul', 'sananduva', 'santa-barbara-do-sul', 'santa-cecilia-do-sul', 'santa-clara-do-sul', 'santa-cruz-do-sul', 'santa-maria', 'santa-maria-do-herval', 'santa-margarida-do-sul', 'santana-da-boa-vista', 'santana-do-livramento', 'santa-rosa', 'santa-tereza', 'santa-vitoria-do-palmar', 'santiago', 'santo-angelo', 'santo-antonio-do-palma', 'santo-antonio-da-patrulha', 'santo-antonio-das-missoes', 'santo-antonio-do-planalto', 'santo-augusto', 'santo-cristo', 'santo-expedito-do-sul', 'sao-borja', 'sao-domingos-do-sul', 'sao-francisco-de-assis', 'sao-francisco-de-paula', 'sao-gabriel', 'sao-jeronimo', 'sao-joao-da-urtiga', 'sao-joao-do-polesine', 'sao-jorge', 'sao-jose-das-missoes', 'sao-jose-do-herval', 'sao-jose-do-hortencio', 'sao-jose-do-inhacora', 'sao-jose-do-norte', 'sao-jose-do-ouro', 'sao-jose-do-sul', 'sao-jose-dos-ausentes', 'sao-leopoldo', 'sao-lourenco-do-sul', 'sao-luiz-gonzaga', 'sao-marcos', 'sao-martinho', 'sao-martinho-da-serra', 'sao-miguel-das-missoes', 'sao-nicolau', 'sao-paulo-das-missoes', 'sao-pedro-da-serra', 'sao-pedro-das-missoes', 'sao-pedro-do-butia', 'sao-pedro-do-sul', 'sao-sebastiao-do-cai', 'sao-sepe', 'sao-valentim', 
-        'sao-valentim-do-sul', 'sao-valerio-do-sul', 'sao-vendelino', 'sao-vicente-do-sul', 'sapiranga', 'sapucaia-do-sul', 'sarandi', 'seberi', 'sede-nova', 'segredo', 'selbach', 'senador-salgado-filho', 'sentinela-do-sul', 'serafina-correa', 'serio', 'sertao', 'sertao-santana', 'sete-de-setembro', 'severiano-de-almeida', 'silveira-martins', 'sinimbu', 'sobradinho', 'soledade', 'tabai', 'tapejara', 'tapera', 'tapes', 'taquara', 'taquari', 'taquarucu-do-sul', 'tavares', 'tenente-portela', 'terra-de-areia', 'teutonia', 'tio-hugo', 'tiradentes-do-sul', 'toropi', 'torres', 'tramandai', 'travesseiro', 'tres-arroios', 'tres-cachoeiras', 'tres-coroas', 'tres-de-maio', 'tres-forquilhas', 'tres-palmeiras', 'tres-passos', 'trindade-do-sul', 'triunfo', 'tucunduva', 'tunas', 'tupanci-do-sul', 'tupancireta', 'tupandi', 'tuparendi', 'turucu', 'ubiretama', 'uniao-da-serra', 'unistalda', 'uruguaiana', 'vacaria', 'vale-verde', 'vale-do-sol', 'vale-real', 'vanini', 'venancio-aires', 'vera-cruz', 'veranopolis', 'vespasiano-correa', 'viadutos', 'viamao', 'vicente-dutra', 'victor-graeff', 'vila-flores', 'vila-langaro', 'vila-maria', 'vila-nova-do-sul', 'vista-alegre', 'vista-alegre-do-prata', 'vista-gaucha', 'vitoria-das-missoes', 'westfalia', 'xangri-la'], 'SC': ['abdon-batista', 'abelardo-luz', 'agrolandia', 'agronomica', 'agua-doce', 'aguas-de-chapeco', 'aguas-frias', 'aguas-mornas', 'alfredo-wagner', 'alto-bela-vista', 'anchieta', 'angelina', 'anita-garibaldi', 'anitapolis', 'antonio-carlos', 'apiuna', 'arabuta', 'araquari', 'ararangua', 'armazem', 'arroio-trinta', 'arvoredo', 'ascurra', 'atalanta', 'aurora', 'balneario-arroio-do-silva', 'balneario-camboriu', 'balneario-barra-do-sul', 'balneario-gaivota', 'bandeirante', 'barra-bonita', 'barra-velha', 'bela-vista-do-toldo', 'belmonte', 'benedito-novo', 'biguacu', 'blumenau', 'bocaina-do-sul', 
-        'bombinhas', 'bom-jardim-da-serra', 'bom-jesus', 'bom-jesus-do-oeste', 'bom-retiro', 'botuvera', 'braco-do-norte', 'braco-do-trombudo', 'brunopolis', 'brusque', 'cacador', 'caibi', 'calmon', 'camboriu', 'capao-alto', 'campo-alegre', 'campo-belo-do-sul', 'campo-ere', 'campos-novos', 'canelinha', 'canoinhas', 'capinzal', 'capivari-de-baixo', 'catanduvas', 'caxambu-do-sul', 'celso-ramos', 'cerro-negro', 'chapadao-do-lageado', 'chapeco', 'cocal-do-sul', 'concordia', 'cordilheira-alta', 'coronel-freitas', 'coronel-martins', 'corupa', 'correia-pinto', 'criciuma', 'cunha-pora', 'cunhatai', 'curitibanos', 'descanso', 'dionisio-cerqueira', 'dona-emma', 'doutor-pedrinho', 'entre-rios', 'ermo', 'erval-velho', 'faxinal-dos-guedes', 'flor-do-sertao', 'florianopolis', 'formosa-do-sul', 'forquilhinha', 'fraiburgo', 'frei-rogerio', 'galvao', 'garopaba', 'garuva', 'gaspar', 'governador-celso-ramos', 'grao-para', 'gravatal', 'guabiruba', 'guaraciaba', 'guaramirim', 'guaruja-do-sul', 'guatambu', 'herval-d-oeste', 'ibiam', 'ibicare', 'ibirama', 'icara', 'ilhota', 'imarui', 'imbituba', 'imbuia', 'indaial', 'iomere', 'ipira', 'ipora-do-oeste', 'ipuacu', 'ipumirim', 'iraceminha', 'irani', 'irati', 'irineopolis', 'ita', 'itaiopolis', 'itajai', 'itapema', 'itapiranga', 'itapoa', 'ituporanga', 'jabora', 'jacinto-machado', 'jaguaruna', 'jaragua-do-sul', 'jardinopolis', 'joacaba', 'joinville', 'jose-boiteux', 'jupia', 'lacerdopolis', 'lages', 'laguna', 'lajeado-grande', 'laurentino', 'lauro-muller', 'lebon-regis', 'leoberto-leal', 'lindoia-do-sul', 'lontras', 'luiz-alves', 'luzerna', 'macieira', 'mafra', 'major-gercino', 'major-vieira', 'maracaja', 'maravilha', 'marema', 'massaranduba', 'matos-costa', 'meleiro', 'mirim-doce', 'modelo', 'mondai', 'monte-carlo', 'monte-castelo', 'morro-da-fumaca', 'morro-grande', 'navegantes', 'nova-erechim', 'nova-itaberaba', 'nova-trento', 'nova-veneza', 'novo-horizonte', 'orleans', 'otacilio-costa', 'ouro', 'ouro-verde', 'paial', 'painel', 'palhoca', 'palma-sola', 'palmeira', 'palmitos', 'papanduva', 'paraiso', 'passo-de-torres', 'passos-maia', 'paulo-lopes', 'pedras-grandes', 'penha', 'peritiba', 'pescaria-brava', 'petrolandia', 'balneario-picarras', 'pinhalzinho', 'pinheiro-preto', 'piratuba', 'planalto-alegre', 'pomerode', 'ponte-alta', 'ponte-alta-do-norte', 'ponte-serrada', 'porto-belo', 'porto-uniao', 'pouso-redondo', 'praia-grande', 'presidente-castello-branco', 'presidente-getulio', 'presidente-nereu', 'princesa', 'quilombo', 'rancho-queimado', 'rio-das-antas', 'rio-do-campo', 'rio-do-oeste', 'rio-dos-cedros', 'rio-do-sul', 'rio-fortuna', 'rio-negrinho', 'rio-rufino', 'riqueza', 'rodeio', 'romelandia', 'salete', 'saltinho', 'salto-veloso', 'sangao', 'santa-cecilia', 'santa-helena', 'santa-rosa-de-lima', 'santa-rosa-do-sul', 'santa-terezinha', 'santa-terezinha-do-progresso', 'santiago-do-sul', 'santo-amaro-da-imperatriz', 'sao-bernardino', 'sao-bento-do-sul', 'sao-bonifacio', 'sao-carlos', 'sao-cristovao-do-sul', 'sao-domingos', 'sao-francisco-do-sul', 'sao-joao-do-oeste', 'sao-joao-batista', 'sao-joao-do-itaperiu', 'sao-joao-do-sul', 'sao-joaquim', 'sao-jose', 'sao-jose-do-cedro', 'sao-jose-do-cerrito', 'sao-lourenco-do-oeste', 'sao-ludgero', 'sao-martinho', 'sao-miguel-da-boa-vista', 'sao-miguel-do-oeste', 'sao-pedro-de-alcantara', 'saudades', 'schroeder', 'seara', 'serra-alta', 'sideropolis', 'sombrio', 'sul-brasil', 'taio', 'tangara', 'tigrinhos', 'tijucas', 'timbe-do-sul', 'timbo', 'timbo-grande', 'tres-barras', 'treviso', 'treze-de-maio', 'treze-tilias', 'trombudo-central', 'tubarao', 'tunapolis', 'turvo', 'uniao-do-oeste', 'urubici', 'urupema', 'urussanga', 'vargeao', 'vargem', 'vargem-bonita', 'vidal-ramos', 'videira', 'vitor-meireles', 'witmarsum', 'xanxere', 'xavantina', 'xaxim', 'zortea', 'balneario-rincao'], 'SE': ['amparo-de-sao-francisco', 'aquidaba', 'aracaju', 'araua', 'areia-branca', 'barra-dos-coqueiros', 'boquim', 'brejo-grande', 'campo-do-brito', 'canhoba', 'caninde-de-sao-francisco', 'capela', 'carira', 'carmopolis', 'cedro-de-sao-joao', 'cristinapolis', 'cumbe', 'divina-pastora', 'estancia', 'feira-nova', 'frei-paulo', 'gararu', 'general-maynard', 'gracho-cardoso', 'ilha-das-flores', 'indiaroba', 'itabaiana', 'itabaianinha', 'itabi', 'itaporanga-d-ajuda', 'japaratuba', 'japoata', 'lagarto', 'laranjeiras', 'macambira', 'malhada-dos-bois', 'malhador', 'maruim', 'moita-bonita', 'monte-alegre-de-sergipe', 'muribeca', 'neopolis', 'nossa-senhora-aparecida', 'nossa-senhora-da-gloria', 'nossa-senhora-das-dores', 'nossa-senhora-de-lourdes', 'nossa-senhora-do-socorro', 'pacatuba', 'pedra-mole', 'pedrinhas', 'pinhao', 'pirambu', 'poco-redondo', 'poco-verde', 'porto-da-folha', 'propria', 'riachao-do-dantas', 'riachuelo', 'ribeiropolis', 'rosario-do-catete', 'salgado', 'santa-luzia-do-itanhy', 'santana-do-sao-francisco', 'santa-rosa-de-lima', 'santo-amaro-das-brotas', 'sao-cristovao', 'sao-domingos', 'sao-francisco', 'sao-miguel-do-aleixo', 'simao-dias', 'siriri', 'telha', 'tobias-barreto', 'tomar-do-geru', 'umbauba'], 'SP': ['embu-das-artes', 'adamantina', 'adolfo', 'aguai', 'aguas-da-prata', 'aguas-de-lindoia', 'aguas-de-santa-barbara', 'aguas-de-sao-pedro', 
-        'agudos', 'alambari', 'alfredo-marcondes', 'altair', 'altinopolis', 'alto-alegre', 'aluminio', 'alvares-florence', 'alvares-machado', 'alvaro-de-carvalho', 'alvinlandia', 'americana', 'americo-brasiliense', 'americo-de-campos', 'amparo', 'analandia', 'andradina', 'angatuba', 'anhembi', 'anhumas', 'aparecida', 'aparecida-d-oeste', 'apiai', 'aracariguama', 'aracatuba', 
-        'aracoiaba-da-serra', 'aramina', 'arandu', 'arapei', 'araraquara', 'araras', 'arco-iris', 'arealva', 'areias', 'areiopolis', 'ariranha', 'artur-nogueira', 'aruja', 'aspasia', 'assis', 'atibaia', 'auriflama', 'avai', 'avanhandava', 'avare', 'bady-bassitt', 'balbinos', 'balsamo', 'bananal', 'barao-de-antonina', 'barbosa', 'bariri', 'barra-bonita', 'barra-do-chapeu', 'barra-do-turvo', 'barretos', 'barrinha', 'barueri', 'bastos', 'batatais', 'bauru', 'bebedouro', 'bento-de-abreu', 'bernardino-de-campos', 'bertioga', 'bilac', 'birigui', 'biritiba-mirim', 'boa-esperanca-do-sul', 'bocaina', 'bofete', 'boituva', 'bom-jesus-dos-perdoes', 'bom-sucesso-de-itarare', 'bora', 'boraceia', 'borborema', 'borebi', 'botucatu', 'braganca-paulista', 'brauna', 'brejo-alegre', 'brodowski', 'brotas', 'buri', 'buritama', 'buritizal', 'cabralia-paulista', 'cabreuva', 'cacapava', 'cachoeira-paulista', 'caconde', 'cafelandia', 'caiabu', 'caieiras', 'caiua', 'cajamar', 'cajati', 'cajobi', 'cajuru', 'campina-do-monte-alegre', 'campinas', 'campo-limpo-paulista', 'campos-do-jordao', 'campos-novos-paulista', 'cananeia', 'canas', 'candido-mota', 'candido-rodrigues', 'canitar', 'capao-bonito', 'capela-do-alto', 'capivari', 'caraguatatuba', 'carapicuiba', 'cardoso', 'casa-branca', 'cassia-dos-coqueiros', 'castilho', 'catanduva', 'catigua', 'cedral', 'cerqueira-cesar', 'cerquilho', 'cesario-lange', 'charqueada', 'clementina', 'colina', 'colombia', 'conchal', 'conchas', 'cordeiropolis', 'coroados', 'coronel-macedo', 'corumbatai', 'cosmopolis', 'cosmorama', 'cotia', 'cravinhos', 'cristais-paulista', 'cruzalia', 'cruzeiro', 'cubatao', 'cunha', 'descalvado', 'diadema', 'dirce-reis', 'divinolandia', 'dobrada', 'dois-corregos', 'dolcinopolis', 'dourado', 'dracena', 'duartina', 'dumont', 'echapora', 'eldorado', 'elias-fausto', 'elisiario', 'embauba', 'embu', 'embu-guacu', 'emilianopolis', 'engenheiro-coelho', 'espirito-santo-do-pinhal', 'espirito-santo-do-turvo', 'estrela-d-oeste', 'estrela-do-norte', 'euclides-da-cunha-paulista', 'fartura', 'fernandopolis', 'fernando-prestes', 'fernao', 'ferraz-de-vasconcelos', 'flora-rica', 'floreal', 'florida-paulista', 'florinia', 'franca', 'francisco-morato', 'franco-da-rocha', 'gabriel-monteiro', 'galia', 'garca', 'gastao-vidigal', 'gaviao-peixoto', 'general-salgado', 'getulina', 'glicerio', 'guaicara', 'guaimbe', 'guaira', 'guapiacu', 'guapiara', 'guara', 'guaracai', 'guaraci', 'guarani-d-oeste', 'guaranta', 'guararapes', 'guararema', 'guaratingueta', 'guarei', 'guariba', 'guaruja', 'guarulhos', 'guatapara', 'guzolandia', 'herculandia', 'holambra', 'hortolandia', 'iacanga', 'iacri', 'iaras', 'ibate', 'ibira', 'ibirarema', 'ibitinga', 'ibiuna', 'icem', 'iepe', 'igaracu-do-tiete', 'igarapava', 'igarata', 'iguape', 'ilhabela', 'ilha-comprida', 'ilha-solteira', 'indaiatuba', 'indiana', 'indiapora', 'inubia-paulista', 'ipaussu', 'ipero', 'ipeuna', 'ipigua', 'iporanga', 'ipua', 'iracemapolis', 'irapua', 'irapuru', 'itabera', 'itai', 'itajobi', 'itaju', 'itanhaem', 'itaoca', 'itapecerica-da-serra', 'itapetininga', 'itapeva', 'itapevi', 'itapira', 'itapirapua-paulista', 'itapolis', 'itaporanga', 'itapui', 'itapura', 'itaquaquecetuba', 'itarare', 'itariri', 'itatiba', 'itatinga', 'itirapina', 'itirapua', 'itobi', 'itu', 'itupeva', 'ituverava', 'jaborandi', 'jaboticabal', 'jacarei', 'jaci', 'jacupiranga', 'jaguariuna', 'jales', 'jambeiro', 'jandira', 'jardinopolis', 'jarinu', 'jau', 'jeriquara', 'joanopolis', 'joao-ramalho', 'jose-bonifacio', 'julio-mesquita', 'jumirim', 'jundiai', 'junqueiropolis', 'juquia', 'juquitiba', 'lagoinha', 'laranjal-paulista', 'lavinia', 'lavrinhas', 'leme', 'lencois-paulista', 'limeira', 'lindoia', 'lins', 'lorena', 
-        'lourdes', 'louveira', 'lucelia', 'lucianopolis', 'luis-antonio', 'luiziania', 'lupercio', 'lutecia', 'macatuba', 'macaubal', 'macedonia', 'magda', 'mairinque', 'mairipora', 'manduri', 
-        'maraba-paulista', 'maracai', 'marapoama', 'mariapolis', 'marilia', 'marinopolis', 'martinopolis', 'matao', 'maua', 'mendonca', 'meridiano', 'mesopolis', 'miguelopolis', 'mineiros-do-tiete', 'miracatu', 'mira-estrela', 'mirandopolis', 'mirante-do-paranapanema', 'mirassol', 'mirassolandia', 'mococa', 'mogi-das-cruzes', 'mogi-guacu', 'mogi-mirim', 'mombuca', 'moncoes', 
-        'mongagua', 'monte-alegre-do-sul', 'monte-alto', 'monte-aprazivel', 'monte-azul-paulista', 'monte-castelo', 'monteiro-lobato', 'monte-mor', 'morro-agudo', 'morungaba', 'motuca', 'murutinga-do-sul', 'nantes', 'narandiba', 'natividade-da-serra', 'nazare-paulista', 'neves-paulista', 'nhandeara', 'nipoa', 'nova-alianca', 'nova-campina', 'nova-canaa-paulista', 'nova-castilho', 'nova-europa', 'nova-granada', 'nova-guataporanga', 'nova-independencia', 'novais', 'nova-luzitania', 'nova-odessa', 'novo-horizonte', 'nuporanga', 'ocaucu', 'oleo', 'olimpia', 'onda-verde', 'oriente', 'orindiuva', 'orlandia', 'osasco', 'oscar-bressane', 'osvaldo-cruz', 'ourinhos', 'ouroeste', 'ouro-verde', 'pacaembu', 'palestina', 'palmares-paulista', 'palmeira-d-oeste', 'palmital', 'panorama', 'paraguacu-paulista', 'paraibuna', 'paraiso', 'paranapanema', 'paranapua', 'parapua', 'pardinho', 'pariquera-acu', 'parisi', 'patrocinio-paulista', 'pauliceia', 'paulinia', 'paulistania', 'paulo-de-faria', 'pederneiras', 'pedra-bela', 'pedranopolis', 'pedregulho', 'pedreira', 'pedrinhas-paulista', 'pedro-de-toledo', 'penapolis', 'pereira-barreto', 'pereiras', 'peruibe', 'piacatu', 'piedade', 'pilar-do-sul', 'pindamonhangaba', 'pindorama', 'pinhalzinho', 'piquerobi', 'piquete', 'piracaia', 'piracicaba', 'piraju', 'pirajui', 'pirangi', 'pirapora-do-bom-jesus', 'pirapozinho', 'pirassununga', 'piratininga', 'pitangueiras', 'planalto', 'platina', 'poa', 'poloni', 'pompeia', 'pongai', 'pontal', 'pontalinda', 'pontes-gestal', 'populina', 'porangaba', 'porto-feliz', 'porto-ferreira', 'potim', 'potirendaba', 'pracinha', 'pradopolis', 'praia-grande', 'pratania', 'presidente-alves', 'presidente-bernardes', 'presidente-epitacio', 'presidente-prudente', 'presidente-venceslau', 'promissao', 'quadra', 'quata', 'queiroz', 'queluz', 'quintana', 'rafard', 'rancharia', 'redencao-da-serra', 'regente-feijo', 'reginopolis', 'registro', 'restinga', 'ribeira', 'ribeirao-bonito', 'ribeirao-branco', 'ribeirao-corrente', 'ribeirao-do-sul', 'ribeirao-dos-indios', 'ribeirao-grande', 'ribeirao-pires', 'ribeirao-preto', 'riversul', 'rifaina', 'rincao', 'rinopolis', 'rio-claro', 'rio-das-pedras', 'rio-grande-da-serra', 'riolandia', 'rosana', 'roseira', 'rubiacea', 'rubineia', 'sabino', 'sagres', 'sales', 'sales-oliveira', 'salesopolis', 'salmourao', 'saltinho', 'salto', 'salto-de-pirapora', 'salto-grande', 'sandovalina', 'santa-adelia', 'santa-albertina', 'santa-barbara-d-oeste', 'santa-branca', 'santa-clara-d-oeste', 'santa-cruz-da-conceicao', 'santa-cruz-da-esperanca', 'santa-cruz-das-palmeiras', 'santa-cruz-do-rio-pardo', 'santa-ernestina', 'santa-fe-do-sul', 'santa-gertrudes', 'santa-isabel', 'santa-lucia', 'santa-maria-da-serra', 'santa-mercedes', 'santana-da-ponte-pensa', 'santana-de-parnaiba', 'santa-rita-d-oeste', 'santa-rita-do-passa-quatro', 'santa-rosa-de-viterbo', 'santa-salete', 'santo-anastacio', 'santo-andre', 'santo-antonio-da-alegria', 'santo-antonio-de-posse', 'santo-antonio-do-aracangua', 'santo-antonio-do-jardim', 'santo-antonio-do-pinhal', 'santo-expedito', 'santopolis-do-aguapei', 'santos', 'sao-bento-do-sapucai', 'sao-bernardo-do-campo', 
-        'sao-caetano-do-sul', 'sao-carlos', 'sao-francisco', 'sao-joao-da-boa-vista', 'sao-joao-das-duas-pontes', 'sao-joao-de-iracema', 'sao-joao-do-pau-d-alho', 'sao-joaquim-da-barra', 'sao-jose-da-bela-vista', 'sao-jose-do-barreiro', 'sao-jose-do-rio-pardo', 'sao-jose-do-rio-preto', 'sao-jose-dos-campos', 'sao-lourenco-da-serra', 'sao-luis-do-paraitinga', 'sao-manuel', 'sao-miguel-arcanjo', 'sao-paulo', 'sao-pedro', 'sao-pedro-do-turvo', 'sao-roque', 'sao-sebastiao', 'sao-sebastiao-da-grama', 'sao-simao', 'sao-vicente', 'sarapui', 'sarutaia', 'sebastianopolis-do-sul', 'serra-azul', 'serrana', 'serra-negra', 'sertaozinho', 'sete-barras', 'severinia', 'silveiras', 'socorro', 'sorocaba', 'sud-mennucci', 'sumare', 'suzano', 'suzanapolis', 
-        'tabapua', 'tabatinga', 'taboao-da-serra', 'taciba', 'taguai', 'taiacu', 'taiuva', 'tambau', 'tanabi', 'tapirai', 'tapiratiba', 'taquaral', 'taquaritinga', 'taquarituba', 'taquarivai', 
-        'tarabai', 'taruma', 'tatui', 'taubate', 'tejupa', 'teodoro-sampaio', 'terra-roxa', 'tiete', 'timburi', 'torre-de-pedra', 'torrinha', 'trabiju', 'tremembe', 'tres-fronteiras', 'tuiuti', 'tupa', 'tupi-paulista', 'turiuba', 'turmalina', 'ubarana', 'ubatuba', 'ubirajara', 'uchoa', 'uniao-paulista', 'urania', 'uru', 'urupes', 'valentim-gentil', 'valinhos', 'valparaiso', 'vargem', 'vargem-grande-do-sul', 'vargem-grande-paulista', 'varzea-paulista', 'vera-cruz', 'vinhedo', 'viradouro', 'vista-alegre-do-alto', 'vitoria-brasil', 'votorantim', 'votuporanga', 'zacarias', 'chavantes', 'estiva-gerbi'], 'TO': ['abreulandia', 'aguiarnopolis', 'alianca-do-tocantins', 'almas', 'alvorada', 'ananas', 'angico', 'aparecida-do-rio-negro', 'aragominas', 'araguacema', 'araguacu', 'araguaina', 'araguana', 'araguatins', 'arapoema', 'arraias', 'augustinopolis', 'aurora-do-tocantins', 'axixa-do-tocantins', 'babaculandia', 'bandeirantes-do-tocantins', 'barra-do-ouro', 'barrolandia', 'bernardo-sayao', 'bom-jesus-do-tocantins', 'brasilandia-do-tocantins', 'brejinho-de-nazare', 'buriti-do-tocantins', 'cachoeirinha', 'campos-lindos', 'cariri-do-tocantins', 'carmolandia', 'carrasco-bonito', 'caseara', 'centenario', 'chapada-de-areia', 'chapada-da-natividade', 'colinas-do-tocantins', 'combinado', 'conceicao-do-tocantins', 'couto-magalhaes', 'cristalandia', 'crixas-do-tocantins', 'darcinopolis', 'dianopolis', 'divinopolis-do-tocantins', 'dois-irmaos-do-tocantins', 'duere', 'esperantina', 'fatima', 'figueiropolis', 'filadelfia', 'formoso-do-araguaia', 'fortaleza-do-tabocao', 'goianorte', 'goiatins', 'guarai', 'gurupi', 'ipueiras', 'itacaja', 'itaguatins', 'itapiratins', 'itapora-do-tocantins', 'jau-do-tocantins', 'juarina', 'lagoa-da-confusao', 'lagoa-do-tocantins', 'lajeado', 'lavandeira', 'lizarda', 'luzinopolis', 'marianopolis-do-tocantins', 'mateiros', 'maurilandia-do-tocantins', 'miracema-do-tocantins', 'miranorte', 'monte-do-carmo', 'monte-santo-do-tocantins', 'palmeiras-do-tocantins', 'muricilandia', 'natividade', 'nazare', 'nova-olinda', 'nova-rosalandia', 'novo-acordo', 'novo-alegre', 'novo-jardim', 'oliveira-de-fatima', 'palmeirante', 'palmeiropolis', 'paraiso-do-tocantins', 'parana', 'pau-d-arco', 'pedro-afonso', 'peixe', 'pequizeiro', 'colmeia', 'pindorama-do-tocantins', 'piraque', 'pium', 'ponte-alta-do-bom-jesus', 'ponte-alta-do-tocantins', 'porto-alegre-do-tocantins', 'porto-nacional', 'praia-norte', 'presidente-kennedy', 'pugmil', 'recursolandia', 'riachinho', 'rio-da-conceicao', 'rio-dos-bois', 'rio-sono', 'sampaio', 'sandolandia', 'santa-fe-do-araguaia', 'santa-maria-do-tocantins', 'santa-rita-do-tocantins', 'santa-rosa-do-tocantins', 'santa-tereza-do-tocantins', 'santa-terezinha-tocantins', 'sao-bento-do-tocantins', 'sao-felix-do-tocantins', 
-        'sao-miguel-do-tocantins', 'sao-salvador-do-tocantins', 'sao-sebastiao-do-tocantins', 'sao-valerio-da-natividade', 'silvanopolis', 'sitio-novo-do-tocantins', 'sucupira', 'taguatinga', 'taipas-do-tocantins', 'talisma', 'palmas', 'tocantinia', 'tocantinopolis', 'tupirama', 'tupiratins', 'wanderlandia', 'xambioa', 'couto-de-magalhaes']}
-    
-        for index, row in df.iterrows():
-            cidade = row["city"]
-            estado = row["state"]
-            print(f"{cidade} - {estado}")
-
-            # Encontrar a cidade mais próxima
-            cidade_correta = find_closest_match(cidade, lista_cidades[estado])
-            df.at[index, "city"] = cidade_correta
-
-            print(f"Nome digitado: {cidade}")
-            print(f"Nome mais próximo: {cidade_correta}")
-
-        # Salvar o DataFrame limpo em um novo arquivo Excel
-        try:
-            df.to_excel(fixed_file, index=False, engine='openpyxl')
-            print(f"✅ Arquivo salvo como: {fixed_file}")
-            xlsx_file=None
-        except Exception as e:
-            print(f"⚠️ Erro ao salvar o arquivo Excel: {e}")
-        
-    else:
-        print("❌ Nenhum arquivo selecionado")
-
-def levenshtein_distance(s1, s2):
-    """Calcula a Distância de Levenshtein entre duas strings."""
-    len_s1, len_s2 = len(s1), len(s2)
-    dp = np.zeros((len_s1 + 1, len_s2 + 1), dtype=int)
-
-    # Inicializa a matriz
-    for i in range(len_s1 + 1):
-        dp[i][0] = i
-    for j in range(len_s2 + 1):
-        dp[0][j] = j
-
-    # Preenche a matriz
-    for i in range(1, len_s1 + 1):
-        for j in range(1, len_s2 + 1):
-            if s1[i - 1] == s2[j - 1]:  
-                cost = 0  
-            else:
-                cost = 1  
-            
-            dp[i][j] = min(dp[i - 1][j] + 1,      # Remoção
-                        dp[i][j - 1] + 1,      # Inserção
-                        dp[i - 1][j - 1] + cost)  # Substituição
-    
-    return dp[len_s1, len_s2]
-
-def find_closest_match(word, word_list):
-    """ Encontrar se a palavra existe na lista"""
-    for cidades in word_list:
-        if word == cidades:
-            print("saida1")
-            return word
-    
-    """Se não encontra na lista, encontra a palavra mais próxima dentro de uma lista usando Levenshtein Distance."""
-    closest_word = None
-    min_distance = float('inf')
-
-    for candidate in word_list:
-        distance = levenshtein_distance(word.lower(), candidate.lower())
-        if distance < min_distance:
-            min_distance = distance
-            closest_word = candidate
-
-    if closest_word.lower() == word.lower():
-        return word
-    else:
-        return closest_word
-
